20813: Migrate react operations (#231)
- Migrates react, react_series, react_group, react_into_features, and
react_aggregate (and related methods)
- Some minor typing updates
- Renamed some references of `core` to `engine`
