+++ howso/client/base.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from collections.abc import Callable, Collection, Iterable, Mapping, MutableMapping
+from collections.abc import Callable, Collection, Iterable, Mapping, MutableMapping, Sized
 from pathlib import Path
 import typing as t
 from uuid import UUID
@@ -22,10 +22,14 @@ from .typing import (
     Cases,
     Distances,
     Evaluation,
+    GenerateNewCases,
     Mode,
+    NewCaseThreshold,
     Persistence,
     Precision,
+    SeriesIDTracking,
     TabularData2D,
+    TabularData3D,
     TargetedModel,
 )
 
@@ -48,7 +52,7 @@ class AbstractHowsoClient(ABC):
     WARNING_MESSAGES = {
         "invalid_precision": (
             'Supported values for `precision` are "exact" and "similar". The operation will be completed as '
-            'if the value of `precision` is "exact".')
+            'if the value of `%s` is "exact".')
     }
     """Mapping of warning type to default warning message."""
 
@@ -143,6 +147,40 @@ class AbstractHowsoClient(ABC):
         """
 
     @abstractmethod
+    def _should_react_batch(self, params: dict, total_size: int) -> bool:
+        """
+        Determine if given react should be batched.
+
+        Parameters
+        ----------
+        params : dict
+            The react request parameters.
+        total_size : int
+            The size of the cases being reacted to.
+
+        Returns
+        -------
+        bool
+            Whether a react should be batched.
+        """
+
+    @abstractmethod
+    def _get_trainee_thread_count(self, trainee_id: str) -> int:
+        """
+        Get the number of available cpu threads a Trainee has access to.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The identifier of the Trainee.
+
+        Returns
+        -------
+        int
+            The allocated number of cpu threads for a Trainee.
+        """
+
+    @abstractmethod
     def execute(self, trainee_id: str, label: str, payload: t.Any, **kwargs) -> t.Any:
         """
         Execute a label in Howso engine.
@@ -150,7 +188,7 @@ class AbstractHowsoClient(ABC):
         Parameters
         ----------
         trainee_id : str
-            The entity handle of the Trainee.
+            The identifier of the Trainee.
         label : str
             The label to execute.
         payload : Any
@@ -170,7 +208,7 @@ class AbstractHowsoClient(ABC):
         Parameters
         ----------
         trainee_id : str
-            The entity handle of the Trainee.
+            The identifier of the Trainee.
         label : str
             The label to execute.
         payload : Any
@@ -611,7 +649,7 @@ class AbstractHowsoClient(ABC):
             raise ValueError('num_cases must be a value greater than 0')
 
         if precision is not None and precision not in self.SUPPORTED_PRECISION_VALUES:
-            warnings.warn(self.WARNING_MESSAGES['invalid_precision'])
+            warnings.warn(self.WARNING_MESSAGES['invalid_precision'].format("precision"))
 
         # Convert session instance to id
         if (
@@ -739,7 +777,7 @@ class AbstractHowsoClient(ABC):
             raise ValueError('num_cases must be a value greater than 0')
 
         if precision is not None and precision not in self.SUPPORTED_PRECISION_VALUES:
-            warnings.warn(self.WARNING_MESSAGES['invalid_precision'])
+            warnings.warn(self.WARNING_MESSAGES['invalid_precision'].format("precision"))
 
         # Convert session instance to id
         if (
@@ -836,7 +874,7 @@ class AbstractHowsoClient(ABC):
             raise HowsoError(self.ERROR_MESSAGES["missing_session"], code="missing_session")
 
         if precision is not None and precision not in self.SUPPORTED_PRECISION_VALUES:
-            warnings.warn(self.WARNING_MESSAGES['invalid_precision'])
+            warnings.warn(self.WARNING_MESSAGES['invalid_precision'].format("precision"))
 
         if case_indices is not None:
             util.validate_case_indices(case_indices)
@@ -878,20 +916,70 @@ class AbstractHowsoClient(ABC):
             return 0
         return result.get('count', 0)
 
-    @abstractmethod
-    def remove_series_store(self, trainee_id, series=None):
-        """Clear stored series from trainee."""
+    def remove_series_store(self, trainee_id: str, series: t.Optional[str] = None):
+        """
+        Clear any stored series from the Trainee.
 
-    @abstractmethod
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to remove the series store from.
+        series : str, optional
+            The ID of the series to clear.
+            If None, the Trainee's entire series store will be cleared.
+        """
+        trainee_id = self._resolve_trainee(trainee_id)
+        if self.configuration.verbose:
+            print(f'Removing stored series from Trainee with id: {trainee_id} and series: {series}')
+        self.execute(trainee_id, "remove_series_store", {"series": series})
+
     def append_to_series_store(
         self,
-        trainee_id,
-        series,
-        contexts,
+        trainee_id: str,
+        series: str,
+        contexts: TabularData2D,
         *,
-        context_features=None
+        context_features: t.Optional[Collection[str]] = None
     ):
-        """Append the specified contexts to a series store."""
+        """
+        Append the specified contexts to a series store.
+
+        For use with train series.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to append to.
+        series : str
+            The name of the series store to append to.
+        contexts : list of list of object or pandas.DataFrame
+            The list of list of context values to append to the series.
+        context_features : Collection of str, optional
+            The feature names corresponding to context values.
+        """
+        trainee_id = self._resolve_trainee(trainee_id)
+        cached_trainee = self.trainee_cache.get(trainee_id)
+
+        util.validate_list_shape(contexts, 2, "contexts", "list of object", allow_none=False)
+
+        if context_features is None:
+            context_features = internals.get_features_from_data(
+                contexts,
+                data_parameter='contexts',
+                features_parameter='context_features'
+            )
+
+        # Preprocess contexts
+        serialized_contexts = serialize_cases(contexts, context_features, cached_trainee.features)
+
+        if self.configuration.verbose:
+            print(f'Appending to series store for Trainee with id: {trainee_id}, and series: {series}')
+
+        self.execute(trainee_id, "append_to_series_store", {
+            "context_features": context_features,
+            "context_values": serialized_contexts,
+            "series": series,
+        })
 
     def set_substitute_feature_values(self, trainee_id: str, substitution_value_map: Mapping):
         """
@@ -1104,7 +1192,7 @@ class AbstractHowsoClient(ABC):
         trainee_id = self._resolve_trainee(trainee_id)
 
         if precision is not None and precision not in self.SUPPORTED_PRECISION_VALUES:
-            warnings.warn(self.WARNING_MESSAGES['invalid_precision'])
+            warnings.warn(self.WARNING_MESSAGES['invalid_precision'].format("precision"))
 
         # Convert session instance to id
         if (
@@ -1126,142 +1214,2121 @@ class AbstractHowsoClient(ABC):
             return dict()
         return stats
 
-    @abstractmethod
-    def react_series(
-        self, trainee_id, *,
-        action_features=None,
-        actions=None,
-        batch_size=None,
-        case_indices=None,
-        contexts=None,
-        context_features=None,
-        continue_series=False,
-        continue_series_features=None,
-        continue_series_values=None,
-        derived_action_features=None,
-        derived_context_features=None,
-        desired_conviction=None,
-        details=None,
-        exclude_novel_nominals_from_uniqueness_check=False,
-        feature_bounds_map=None,
-        final_time_steps=None,
-        generate_new_cases="no",
-        init_time_steps=None,
-        initial_batch_size=None,
-        initial_features=None,
-        initial_values=None,
-        input_is_substituted=False,
-        leave_case_out=None,
-        max_series_lengths=None,
-        new_case_threshold="min",
-        num_series_to_generate=1,
-        ordered_by_specified_features=False,
-        output_new_series_ids=True,
-        preserve_feature_values=None,
-        progress_callback=None,
-        series_context_features=None,
-        series_context_values=None,
-        series_id_tracking="fixed",
-        series_stop_maps=None,
-        series_index=None,
-        substitute_output=True,
-        suppress_warning=False,
-        use_case_weights=False,
-        use_regional_model_residuals=True,
-        weight_feature=None
+    def react(  # noqa: C901
+        self,
+        trainee_id: str,
+        *,
+        action_features: t.Optional[Collection[str]] = None,
+        actions: t.Optional[TabularData2D] = None,
+        allow_nulls: bool = False,
+        batch_size: t.Optional[int] = None,
+        case_indices: t.Optional[CaseIndices] = None,
+        contexts: t.Optional[TabularData2D] = None,
+        context_features: t.Optional[Collection[str]] = None,
+        derived_action_features: t.Optional[Collection[str]] = None,
+        derived_context_features: t.Optional[Collection[str]] = None,
+        desired_conviction: t.Optional[float] = None,
+        details: t.Optional[Mapping] = None,
+        exclude_novel_nominals_from_uniqueness_check: bool = False,
+        feature_bounds_map: t.Optional[Mapping] = None,
+        generate_new_cases: GenerateNewCases = "no",
+        initial_batch_size: t.Optional[int] = None,
+        input_is_substituted: bool = False,
+        into_series_store: t.Optional[str] = None,
+        leave_case_out: bool = False,
+        new_case_threshold: NewCaseThreshold = "min",
+        num_cases_to_generate: int = 1,
+        ordered_by_specified_features: bool = False,
+        post_process_features: t.Optional[Collection[str]] = None,
+        post_process_values: t.Optional[TabularData2D] = None,
+        preserve_feature_values: t.Optional[Collection[str]] = None,
+        progress_callback: t.Optional[Callable] = None,
+        substitute_output: bool = True,
+        suppress_warning: bool = False,
+        use_case_weights: bool = False,
+        use_regional_model_residuals: bool = True,
+        weight_feature: t.Optional[str] = None,
     ) -> Reaction:
-        """React in a series until a stop condition is met."""
+        r"""
+        React to supplied values and cases contained within the Trainee.
 
-    @abstractmethod
+        If desired_conviction is not specified, executes a discriminative
+        react: provided a list of context values, the trainee reacts to the
+        model and produces predictions for the specified actions. If
+        desired_conviction is specified, executes a generative react,
+        produces action_values for the specified action_features conditioned
+        on the optionally provided contexts.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to react to.
+
+        contexts : list of list of object or DataFrame, optional
+            A 2d list of context values to react to.
+            If None for discriminative react, it is assumed that `session`
+            and `session_id` keys are set in the `details`.
+
+            >>> contexts = [[1, 2, 3], [4, 5, 6]]
+
+        action_features : iterable of str, optional
+            An iterable of feature names to treat as action features during
+            react.
+
+            >>> action_features = ['rain_chance', 'is_sunny']
+
+        actions : list of list of object or DataFrame, optional
+            One or more action values to use for action features.
+            If specified, will only return the specified explanation
+            details for the given actions. (Discriminative reacts only)
+
+            >>> actions = [[1, 2, 3], [4, 5, 6]]
+
+        allow_nulls : bool, default False
+            When true will allow return of null values if there
+            are nulls in the local model for the action features, applicable
+            only to discriminative reacts.
+
+        batch_size: int, optional
+            Define the number of cases to react to at once. If left unspecified,
+            the batch size will be determined automatically.
+
+        context_features : iterable of str, optional
+            An iterable of feature names to treat as context features during
+            react.
+
+            >>> context_features = ['temperature', 'humidity', 'dew_point',
+            ...                     'barometric_pressure']
+        derived_context_features : iterable of str, optional
+            An iterable of feature names whose values should be computed
+            from the provided context in the specified order. Must be different
+            than context_features.
+        derived_action_features : iterable of str, optional
+            An iterable of feature names whose values should be computed
+            after generation from the generated case prior to output, in the
+            specified order. Must be a subset of action_features.
+
+            .. note::
+
+                Both of these derived feature lists rely on the features'
+                "derived_feature_code" attribute to compute the values. If
+                'derived_feature_code' attribute is undefined or references
+                non-0 feature indices, the derived value will be null.
+
+        input_is_substituted : bool, default False
+            if True assumes provided categorical (nominal or
+            ordinal) feature values have already been substituted.
+        substitute_output : bool, default True
+            If False, will not substitute categorical feature
+            values. Only applicable if a substitution value map has been set.
+        details : dict, optional
+            If details are specified, the response will contain the requested
+            explanation data along with the reaction. Below are the valid keys
+            and data types for the different audit details. Omitted keys,
+            values set to None, or False values for Booleans will not be
+            included in the audit data returned.
+
+            - boundary_cases : bool, optional
+                If True, outputs an automatically determined (when
+                'num_boundary_cases' is not specified) relevant number of
+                boundary cases. Uses both context and action features of the
+                reacted case to determine the counterfactual boundary based on
+                action features, which maximize the dissimilarity of action
+                features while maximizing the similarity of context features.
+                If action features aren't specified, uses familiarity conviction
+                to determine the boundary instead.
+            - boundary_cases_familiarity_convictions : bool, optional
+                If True, outputs familiarity conviction of addition for each of
+                the boundary cases.
+            - case_contributions_full : bool, optional
+                If true outputs each influential case's differences between the
+                predicted action feature value and the predicted action feature
+                value if each individual case were not included. Uses only the
+                context features of the reacted case to determine that area.
+                Uses full calculations, which uses leave-one-out for cases for
+                computations.
+            - case_contributions_robust : bool, optional
+                If true outputs each influential case's differences between the
+                predicted action feature value and the predicted action feature
+                value if each individual case were not included. Uses only the
+                context features of the reacted case to determine that area.
+                Uses robust calculations, which uses uniform sampling from
+                the power set of all combinations of cases.
+            - case_feature_residuals_full : bool, optional
+                If True, outputs feature residuals for all (context and action)
+                features for just the specified case. Uses leave-one-out for
+                each feature, while using the others to predict the left out
+                feature with their corresponding values from this case. Uses
+                full calculations, which uses leave-one-out for cases for
+                computations.
+            - case_feature_residuals_robust : bool, optional
+                If True, outputs feature residuals for all (context and action)
+                features for just the specified case. Uses leave-one-out for
+                each feature, while using the others to predict the left out
+                feature with their corresponding values from this case. Uses
+                robust calculations, which uses uniform sampling from the power
+                set of features as the contexts for predictions.
+            - case_mda_robust : bool, optional
+                If True, outputs each influential case's mean decrease in
+                accuracy of predicting the action feature in the local model
+                area, as if each individual case were included versus not
+                included. Uses only the context features of the reacted case to
+                determine that area. Uses robust calculations, which uses
+                uniform sampling from the power set of all combinations of cases.
+            - case_mda_full : bool, optional
+                If True, outputs each influential case's mean decrease in
+                accuracy of predicting the action feature in the local model
+                area, as if each individual case were included versus not
+                included. Uses only the context features of the reacted case to
+                determine that area. Uses full calculations, which uses
+                leave-one-out for cases for  computations.
+            - categorical_action_probabilities : bool, optional
+                If True, outputs probabilities for each class for the action.
+                Applicable only to categorical action features.
+            - derivation_parameters : bool, optional
+                If True, outputs a dictionary of the parameters used in the
+                react call. These include k, p, distance_transform,
+                feature_weights, feature_deviations, nominal_class_counts,
+                and use_irw.
+
+                - k: the number of cases used for the local model.
+                - p: the parameter for the Lebesgue space.
+                - distance_transform: the distance transform used as an
+                  exponent to convert distances to raw influence weights.
+                - feature_weights: the weight for each feature used in the
+                  distance metric.
+                - feature_deviations: the deviation for each feature used in
+                  the distance metric.
+                - nominal_class_counts: the number of unique values for each
+                  nominal feature. This is used in the distance metric.
+                - use_irw: a flag indicating if feature weights were
+                  derived using inverse residual weighting.
+            - distance_contribution : bool, optional
+                If True, outputs the distance contribution (expected total
+                surprisal contribution) for the reacted case. Uses both context
+                and action feature values.
+            - distance_ratio : bool, optional
+                If True, outputs the ratio of distance (relative surprisal)
+                between this reacted case and its nearest case to the minimum
+                distance (relative surprisal) in between the closest two cases
+                in the local area. All distances are computed using only the
+                specified context features.
+            - feature_contributions_robust : bool, optional
+                If True outputs each context feature's absolute and directional
+                differences between the predicted action feature value and the
+                predicted action feature value if each context were not in the
+                model for all context features in the local model area Uses
+                robust calculations, which uses uniform sampling from the power
+                set of features as the contexts for predictions. Directional feature
+                contributions are returned under the key
+                'directional_feature_contributions_robust'.
+            - feature_contributions_full : bool, optional
+                If True outputs each context feature's absolute and directional
+                differences between the predicted action feature value and the
+                predicted action feature value if each context were not in the
+                model for all context features in the local model area. Uses
+                full calculations, which uses leave-one-out for cases for
+                computations. Directional feature contributions are returned
+                under the key 'directional_feature_contributions_full'.
+            - case_feature_contributions_robust: bool, optional
+                If True outputs each context feature's absolute and directional
+                differences between the predicted action feature value and the
+                predicted action feature value if each context feature were not
+                in the model for all context features in this case, using only
+                the values from this specific case. Uses
+                robust calculations, which uses uniform sampling from the power
+                set of features as the contexts for predictions.
+                Directional case feature contributions are returned under the
+                'case_directional_feature_contributions_robust' key.
+            - case_feature_contributions_full: bool, optional
+                If True outputs each context feature's absolute and directional
+                differences between the predicted action feature value and the
+                predicted action feature value if each context feature were not
+                in the model for all context features in this case, using only
+                the values from this specific case. Uses
+                full calculations, which uses leave-one-out for cases for
+                computations. Directional case feature
+                contributions are returned under the
+                'case_directional_feature_contributions_full' key.
+            - feature_mda_robust : bool, optional
+                If True, outputs each context feature's mean decrease in
+                accuracy of predicting the action feature given the context.
+                Uses only the context features of the reacted case to determine
+                that area. Uses robust calculations, which uses uniform sampling
+                from the power set of features as the contexts for predictions.
+            - feature_mda_full : bool, optional
+                If True, outputs each context feature's mean decrease in
+                accuracy of predicting the action feature given the context.
+                Uses only the context features of the reacted case to determine
+                that area. Uses full calculations, which uses leave-one-out
+                for cases for computations.
+            - feature_mda_ex_post_robust : bool, optional
+                If True, outputs each context feature's mean decrease in
+                accuracy of predicting the action feature as an explanation detail
+                given that the specified prediction was already made as
+                specified by the action value. Uses both context and action
+                features of the reacted case to determine that area. Uses
+                robust calculations, which uses uniform sampling
+                from the power set of features as the contexts for predictions.
+            - feature_mda_ex_post_full : bool, optional
+                If True, outputs each context feature's mean decrease in
+                accuracy of predicting the action feature as an explanation detail
+                given that the specified prediction was already made as
+                specified by the action value. Uses both context and action
+                features of the reacted case to determine that area. Uses
+                full calculations, which uses leave-one-out for cases for
+                computations.
+            - features : list of str, optional
+                A list of feature names that specifies for what features will
+                per-feature details be computed (residuals, contributions,
+                mda, etc.). This should generally preserve compute, but will
+                not when computing details robustly. Details will be computed
+                for all context and action features if this value is not
+                specified.
+            - feature_residual_robust : bool, optional
+                If True, outputs feature residuals for all (context and action)
+                features locally around the prediction. Uses only the context
+                features of the reacted case to determine that area. Uses robust
+                calculations, which uses uniform sampling
+                from the power set of features as the contexts for predictions.
+            - feature_residuals_full : bool, optional
+                If True, outputs feature residuals for all (context and action)
+                features locally around the prediction. Uses only the context
+                features of the reacted case to determine that area. Uses
+                full calculations, which uses leave-one-out for cases for computations.
+            - global_case_feature_residual_convictions_robust : bool, optional
+                If True, outputs this case's feature residual convictions for
+                the global model. Computed as: global model feature residual
+                divided by case feature residual. Uses robust calculations, which
+                uses uniform sampling from the power set of features as the
+                contexts for predictions.
+            - global_case_feature_residual_convictions_full : bool, optional
+                If True, outputs this case's feature residual convictions for
+                the global model. Computed as: global model feature residual
+                divided by case feature residual. Uses full calculations,
+                which uses leave-one-out for cases for computations.
+            - hypothetical_values : dict, optional
+                A dictionary of feature name to feature value. If specified,
+                shows how a prediction could change in a what-if scenario where
+                the influential cases' context feature values are replaced with
+                the specified values.  Iterates over all influential cases,
+                predicting the action features each one using the updated
+                hypothetical values. Outputs the predicted arithmetic over the
+                influential cases for each action feature.
+            - influential_cases : bool, optional
+                If True, outputs the most influential cases and their influence
+                weights based on the surprisal of each case relative to the
+                context being predicted among the cases. Uses only the context
+                features of the reacted case.
+            - influential_cases_familiarity_convictions :  bool, optional
+                If True, outputs familiarity conviction of addition for each of
+                the influential cases.
+            - influential_cases_raw_weights : bool, optional
+                If True, outputs the surprisal for each of the influential
+                cases.
+            - local_case_feature_residual_convictions_robust : bool, optional
+                If True, outputs this case's feature residual convictions for
+                the region around the prediction. Uses only the context
+                features of the reacted case to determine that region.
+                Computed as: region feature residual divided by case feature
+                residual. Uses robust calculations, which uses uniform sampling
+                from the power set of features as the contexts for predictions.
+            - local_case_feature_residual_convictions_full : bool, optional
+                If True, outputs this case's feature residual convictions for
+                the region around the prediction. Uses only the context
+                features of the reacted case to determine that region.
+                Computed as: region feature residual divided by case feature
+                residual. Uses full calculations, which uses leave-one-out
+                for cases for computations.
+            - most_similar_cases : bool, optional
+                If True, outputs an automatically determined (when
+                'num_most_similar_cases' is not specified) relevant number of
+                similar cases, which will first include the influential cases.
+                Uses only the context features of the reacted case.
+            - num_boundary_cases : int, optional
+                Outputs this manually specified number of boundary cases.
+            - num_most_similar_cases : int, optional
+                Outputs this manually specified number of most similar cases,
+                which will first include the influential cases.
+            - num_most_similar_case_indices : int, optional
+                Outputs this specified number of most similar case indices when
+                'distance_ratio' is also set to True.
+            - num_robust_influence_samples_per_case : int, optional
+                Specifies the number of robust samples to use for each case.
+                Applicable only for computing robust feature contributions or
+                robust case feature contributions. Defaults to 2000. Higher
+                values will take longer but provide more stable results.
+            - observational_errors : bool, optional
+                If True, outputs observational errors for all features as
+                defined in feature attributes.
+            - outlying_feature_values : bool, optional
+                If True, outputs the reacted case's context feature values that
+                are outside the min or max of the corresponding feature values
+                of all the cases in the local model area. Uses only the context
+                features of the reacted case to determine that area.
+            - prediction_stats : bool, optional
+                When true outputs feature prediction stats for all (context
+                and action) features locally around the prediction. The stats
+                returned  are ("r2", "rmse", "spearman_coeff", "precision",
+                "recall", "accuracy", "mcc", "confusion_matrix", "missing_value_accuracy").
+                Uses only the context features of the reacted case to determine that area.
+                Uses full calculations, which uses leave-one-out context features for
+                computations.
+            - selected_prediction_stats : list, optional. List of stats to output. When unspecified,
+                returns all except the confusion matrix. Allowed values:
+
+                - all : Returns all the the available prediction stats, including the confusion matrix.
+                - accuracy : The number of correct predictions divided by the
+                  total number of predictions.
+                - confusion_matrix : A sparse map of actual feature value to a map of
+                  predicted feature value to counts.
+                - mae : Mean absolute error. For continuous features, this is
+                  calculated as the mean of absolute values of the difference
+                  between the actual and predicted values. For nominal features,
+                  this is 1 - the average categorical action probability of each case's
+                  correct classes. Categorical action probabilities are the probabilities
+                  for each class for the action feature.
+                - mda : Mean decrease in accuracy when each feature is dropped
+                  from the model, applies to all features.
+                - feature_mda_permutation_full : Mean decrease in accuracy that used
+                  scrambling of feature values instead of dropping each
+                  feature, applies to all features.
+                - precision : Precision (positive predictive) value for nominal
+                  features only.
+                - r2 : The r-squared coefficient of determination, for
+                  continuous features only.
+                - recall : Recall (sensitivity) value for nominal features only.
+                - rmse : Root mean squared error, for continuous features only.
+                - spearman_coeff : Spearman's rank correlation coefficient,
+                  for continuous features only.
+                - mcc : Matthews correlation coefficient, for nominal features only.
+            - similarity_conviction : bool, optional
+                If True, outputs similarity conviction for the reacted case.
+                Uses both context and action feature values as the case values
+                for all computations. This is defined as expected (local)
+                distance contribution divided by reacted case distance
+                contribution.
+            - generate_attempts : bool, optional
+                If True outputs the number of attempts taken to generate each
+                case. Only applicable when 'generate_new_cases' is "always" or
+                "attempt".
+
+            >>> details = {'num_most_similar_cases': 5,
+            ...            'feature_residuals_full': True}
+
+        desired_conviction : float
+            If specified will execute a generative react. If not
+            specified will executed a discriminative react. Conviction is the
+            ratio of expected surprisal to generated surprisal for each
+            feature generated, valid values are in the range of
+            :math:`(0, \\infty)`.
+        weight_feature : str, optional
+            Name of feature whose values to use as case weights.
+            When left unspecified uses the internally managed case weight.
+        use_case_weights : bool, default False
+            If set to True will scale influence weights by each
+            case's weight_feature weight.
+        case_indices : Iterable of Sequence[Union[str, int]], defaults to None
+            An Iterable of Sequences, of session id and index, where
+            index is the original 0-based index of the case as it was trained
+            into the session. If this case does not exist, discriminative react
+            outputs null, generative react ignores it.
+        preserve_feature_values : iterable of str
+            List of features that will preserve their values from
+            the case specified by case_indices, appending and overwriting the
+            specified contexts as necessary.  For generative reacts, if
+            case_indices isn't specified will preserve feature values of a
+            random case.
+        leave_case_out : bool, default False
+            If set to True and specified along with case_indices,
+            each individual react will respectively ignore the corresponding
+            case specified by case_indices by leaving it out.
+        initial_batch_size: int, optional
+            Define the number of cases to react to in the first batch. If
+            unspecified, the value of the ``react_initial_batch_size`` property
+            is used. The number of cases in following batches will be
+            automatically adjusted. This value is ignored if ``batch_size`` is
+            specified.
+        into_series_store : str, optional
+            The name of a series store. If specified, will store an internal
+            record of all react contexts for this session and series to be used
+            later with train series.
+        use_regional_model_residuals : bool
+            If false uses model feature residuals, if True
+            recalculates regional model residuals.
+        feature_bounds_map : dict of dict
+            A mapping of feature names to the bounds for the
+            feature values to be generated in. For continuous features this
+            should be a numeric value, for datetimes this should be a datetime
+            string. Min bounds should be equal to or smaller than max bounds,
+            except when setting the bounds around the cycle length of a cyclic
+            feature.(e.g., to allow 0 +/- 60 degrees, set min=300 and max=60).
+
+            .. code-block::
+                :caption: Example feature bounds map:
+
+                {
+                    "feature_a": {"min": 0},
+                    "feature_b" : {"min": 1, "max": 5},
+                    "feature_c": {"max": 1}
+                }
+
+        generate_new_cases : {"always", "attempt", "no"}, default "no"
+            (Optional) Whether to generate new cases.
+
+            This parameter takes in a string equal to one of the following:
+
+            a. "attempt"
+
+                `Synthesizer` attempts to generate new cases and
+                if its not possible to generate a new case, it might
+                generate cases in "no" mode (see point c.)
+            b. "always"
+
+                `Synthesizer` always generates new cases and
+                if its not possible to generate a new case, it returns
+                `None`.
+            c. "no"
+
+                `Synthesizer` generates data based on the
+                `desired_conviction` specified and the generated data is
+                not guaranteed to be a new case (that is, a case not found
+                in original dataset.)
+
+        ordered_by_specified_features : bool, default False
+            If True order of generated feature values will match
+            the order of specified features.
+        num_cases_to_generate : int, default 1
+            The number of cases to generate.
+        suppress_warning : bool, defaults to False
+            If True, warnings will not be displayed.
+        post_process_features : iterable of str, optional
+            List of feature names that will be made available during the
+            execution of post_process feature attributes.
+        post_process_values : list of list of object or DataFrame, optional
+            A 2d list of values corresponding to post_process_features that
+            will be made available during the execution of post_process feature
+            attributes.
+        progress_callback : callable, optional
+            A callback method that will be called before each
+            batched call to react and at the end of reacting. The method is
+            given a ProgressTimer containing metrics on the progress and timing
+            of the react operation, and the batch result.
+        new_case_threshold : str, optional
+            Distance to determine the privacy cutoff. If None,
+            will default to "min".
+
+            Possible values:
+
+                - min: minimum distance in the original local space.
+                - max: maximum distance in the original local space.
+                - most_similar: distance between the nearest neighbor to the
+                  nearest neighbor in the original space.
+        exclude_novel_nominals_from_uniqueness_check : bool, default False
+            If True, will exclude features which have a subtype defined in their feature
+            attributes from the uniqueness check that happens when ``generate_new_cases``
+            is True. Only applies to generative reacts.
+
+        Returns
+        -------
+        Reaction:
+            A MutableMapping (dict-like) with these keys -> values:
+                action -> pandas.DataFrame
+                    A data frame of action values.
+
+                details -> Dict or List
+                    An aggregated list of any requested details.
+
+        Raises
+        ------
+        ValueError
+            If `derived_action_features` is not a subset of `action_features`.
+
+            If `new_case_threshold` is not one of {"max", "min", "most_similar"}.
+
+            If the number of context values does not match the number of context features.
+        HowsoError
+            If `num_cases_to_generate` is not an integer greater than 0.
+        """
+        trainee_id = self._resolve_trainee(trainee_id)
+        trainee = self.trainee_cache.get(trainee_id)
+        action_features, actions, context_features, contexts = (
+            self._preprocess_react_parameters(
+                action_features=action_features,
+                action_values=actions,
+                case_indices=case_indices,
+                context_features=context_features,
+                context_values=contexts,
+                desired_conviction=desired_conviction,
+                preserve_feature_values=preserve_feature_values,
+                trainee_id=trainee_id
+            )
+        )
+
+        if post_process_values is not None and post_process_features is None:
+            post_process_features = internals.get_features_from_data(
+                post_process_values,
+                data_parameter='post_process_values',
+                features_parameter='post_process_features')
+        post_process_values = serialize_cases(post_process_values, post_process_features, trainee.features)
+
+        if post_process_values is not None and contexts is not None:
+            if (len(contexts) > 1 and len(post_process_values) > 1 and
+                    len(contexts) != len(post_process_values)):
+                raise ValueError(
+                    "If more than one value is provided for 'contexts' "
+                    "and 'post_process_values', then they must be of the same "
+                    "length."
+                )
+
+        if action_features is not None and derived_action_features is not None:
+            if not set(derived_action_features).issubset(set(action_features)):
+                raise ValueError(
+                    'Specified \'derived_action_features\' must be a subset of '
+                    '\'action_features\'.')
+
+        if new_case_threshold not in [None, "min", "max", "most_similar"]:
+            raise ValueError(
+                f"The value '{new_case_threshold}' specified for the parameter "
+                "`new_case_threshold` is not valid. It accepts one of the"
+                " following values - ['min', 'max', 'most_similar',]"
+            )
+
+        if details is not None and 'robust_computation' in details:
+            details = dict(details)
+            details['robust_influences'] = details['robust_computation']
+            details['robust_residuals'] = details['robust_computation']
+            del details['robust_computation']
+            warnings.warn(
+                'The detail "robust_computation" is deprecated and will be '
+                'removed in a future release. Please use "robust_residuals" '
+                'and/or "robust_influences" instead.', DeprecationWarning)
+
+        if desired_conviction is None:
+            if contexts is not None:
+                for context in contexts:
+                    if context is not None and len(context) != len(context_features or []):
+                        raise ValueError(
+                            "The number of provided context values in "
+                            "`contexts` does not match the number of features "
+                            "in `context_features`."
+                        )
+                total_size = len(contexts)
+            elif case_indices is not None:
+                total_size = len(case_indices)
+            else:
+                total_size = 1
+
+            # discriminative react parameters
+            react_params = {
+                "action_values": actions,
+                "context_features": context_features,
+                "context_values": contexts,
+                "action_features": action_features,
+                "derived_context_features": derived_context_features,
+                "derived_action_features": derived_action_features,
+                "post_process_features": post_process_features,
+                "post_process_values": post_process_values,
+                "case_indices": case_indices,
+                "allow_nulls": allow_nulls,
+                "input_is_substituted": input_is_substituted,
+                "substitute_output": substitute_output,
+                "weight_feature": weight_feature,
+                "use_case_weights": use_case_weights,
+                "leave_case_out": leave_case_out,
+                "preserve_feature_values": preserve_feature_values,
+                "new_case_threshold": new_case_threshold,
+                "details": details,
+            }
+        else:
+            if (
+                not isinstance(num_cases_to_generate, int) or
+                num_cases_to_generate <= 0
+            ):
+                raise HowsoError("`num_cases_to_generate` must be an integer greater than 0.")
+            total_size = num_cases_to_generate
+
+            context_features, contexts = \
+                self._preprocess_generate_parameters(
+                    trainee_id,
+                    desired_conviction=desired_conviction,
+                    num_cases_to_generate=num_cases_to_generate,
+                    action_features=action_features,
+                    context_features=context_features,
+                    context_values=contexts,
+                    case_indices=case_indices
+                )
+
+            # generative react parameters
+            react_params = {
+                "num_cases_to_generate": num_cases_to_generate,
+                "context_features": context_features,
+                "context_values": contexts,
+                "action_features": action_features,
+                "derived_context_features": derived_context_features,
+                "derived_action_features": derived_action_features,
+                "post_process_features": post_process_features,
+                "post_process_values": post_process_values,
+                "use_regional_model_residuals": use_regional_model_residuals,
+                "desired_conviction": desired_conviction,
+                "feature_bounds_map": feature_bounds_map,
+                "generate_new_cases": generate_new_cases,
+                "ordered_by_specified_features": ordered_by_specified_features,
+                "preserve_feature_values": preserve_feature_values,
+                "new_case_threshold": new_case_threshold,
+                "into_series_store": into_series_store,
+                "input_is_substituted": input_is_substituted,
+                "substitute_output": substitute_output,
+                "weight_feature": weight_feature,
+                "use_case_weights": use_case_weights,
+                "case_indices": case_indices,
+                "leave_case_out": leave_case_out,
+                "details": details,
+                "exclude_novel_nominals_from_uniqueness_check": exclude_novel_nominals_from_uniqueness_check,
+            }
+
+        if batch_size or self._should_react_batch(react_params, total_size):
+            # Run in batch
+            if self.configuration.verbose:
+                print(f'Batch reacting to context on Trainee with id: {trainee_id}')
+            response = self._batch_react(
+                trainee_id,
+                react_params,
+                batch_size=batch_size,
+                initial_batch_size=initial_batch_size,
+                total_size=total_size,
+                progress_callback=progress_callback
+            )
+        else:
+            # Run as a single react request
+            if self.configuration.verbose:
+                print(f'Reacting to context on Trainee with id: {trainee_id}')
+            with ProgressTimer(total_size) as progress:
+                if isinstance(progress_callback, Callable):
+                    progress_callback(progress, None)
+                response, _, _ = self._react(trainee_id, react_params)
+                progress.update(total_size)
+
+            if isinstance(progress_callback, Callable):
+                progress_callback(progress, response)
+
+        response = internals.format_react_response(response)
+
+        if desired_conviction is not None:
+            # If the number of cases generated is less then requested, raise
+            # warning (only for generative reacts)
+            internals.insufficient_generation_check(
+                num_cases_to_generate, len(response['action']),
+                suppress_warning=suppress_warning
+            )
+
+        return Reaction(response.get('action'), response.get('details'))
+
+    def _batch_react(
+        self,
+        trainee_id: str,
+        params: dict,
+        *,
+        batch_size: t.Optional[int] = None,
+        initial_batch_size: t.Optional[int] = None,
+        total_size: int,
+        progress_callback: t.Optional[Callable] = None
+    ):
+        """
+        Make react requests in batch.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to react to.
+        params : dict
+            The engine react parameters.
+        batch_size : int, optional
+            Define the number of cases to react to at once. If left unspecified,
+            the batch size will be determined automatically.
+        initial_batch_size : int, optional
+            Define the number of cases to react to in the first batch. If
+            unspecified, the value of the ``react_initial_batch_size`` property
+            is used. The number of cases in following batches will be
+            automatically adjusted. This value is ignored if ``batch_size`` is
+            specified.
+        total_size : int
+            The total size of the data that will be batched.
+        progress_callback : callable, optional
+            A method to be called during batching to retrieve the progress
+            metrics.
+
+        Returns
+        -------
+        dict
+            The react response.
+        """
+        temp_result = None
+        accumulated_result = {'action_values': []}
+
+        actions = params.get('action_values')
+        contexts = params.get('context_values')
+        case_indices = params.get('case_indices')
+        post_process_values = params.get('post_process_values')
+
+        with ProgressTimer(total_size) as progress:
+            gen_batch_size = None
+            batch_scaler = None
+            if not batch_size:
+                # Scale the batch size automatically
+                start_batch_size = initial_batch_size or self.react_initial_batch_size
+                batch_scaler = self.batch_scaler_class(start_batch_size, progress)
+                gen_batch_size = batch_scaler.gen_batch_size()
+                batch_size = next(gen_batch_size, None)
+
+            while not progress.is_complete and batch_size is not None:
+                if isinstance(progress_callback, Callable):
+                    progress_callback(progress, temp_result)
+                batch_start = progress.current_tick
+                batch_end = progress.current_tick + batch_size
+
+                if actions is not None and len(actions) > 1:
+                    params['action_values'] = actions[batch_start:batch_end]
+                if contexts is not None and len(contexts) > 1:
+                    params['context_values'] = contexts[batch_start:batch_end]
+                if case_indices is not None and len(case_indices) > 1:
+                    params['case_indices'] = case_indices[batch_start:batch_end]
+                if post_process_values is not None and len(post_process_values) > 1:
+                    params['post_process_values'] = post_process_values[batch_start:batch_end]
+
+                if params.get('desired_conviction') is not None:
+                    params['num_cases_to_generate'] = batch_size
+                temp_result, in_size, out_size = self._react(trainee_id, params)
+
+                internals.accumulate_react_result(accumulated_result, temp_result)
+                if batch_scaler is None or gen_batch_size is None:
+                    progress.update(batch_size)
+                else:
+                    batch_size = batch_scaler.send(
+                        gen_batch_size,
+                        batch_scaler.SendOptions(None, (in_size, out_size)))
+
+        # Final call to callback on completion
+        if isinstance(progress_callback, Callable):
+            progress_callback(progress, temp_result)
+
+        return accumulated_result
+
+    def _react(self, trainee_id: str, params: dict) -> tuple[dict, int, int]:
+        """
+        Make a single react request.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to react to.
+        params : dict
+            The engine react parameters.
+
+        Returns
+        -------
+        dict
+            The react response.
+        int
+            The request payload size.
+        int
+            The response payload size.
+        """
+        ret, in_size, out_size = self.execute_sized(trainee_id, 'react', params)
+
+        # Action values and features should always be defined
+        if not ret.get('action_values'):
+            ret['action_values'] = []
+        if not ret.get('action_features'):
+            ret['action_features'] = []
+
+        return ret, in_size, out_size
+
+    def _preprocess_react_parameters(
+        self,
+        trainee_id: str,
+        *,
+        action_features: t.Optional[Collection[str]],
+        action_values: t.Optional[TabularData2D],
+        case_indices: t.Optional[CaseIndices],
+        context_features: t.Optional[Collection[str]],
+        context_values: t.Optional[TabularData2D],
+        desired_conviction: t.Optional[float],
+        preserve_feature_values: t.Optional[Collection[str]],
+        continue_series: bool = False,
+    ) -> tuple[Collection[str] | None, list[t.Any] | None, Collection[str] | None, list[t.Any] | None]:
+        """
+        Preprocess parameters for `react_` methods.
+
+        Helper method to pre-process the parameters. Updates the passed-in
+        parameters as necessary.
+
+        Parameters
+        ----------
+        trainee_id : str
+            See :meth:`AbstractHowsoClient.react()`.
+        action_features : Collection of str, optional
+            See :meth:`AbstractHowsoClient.react()`.
+        action_values : Collection of object, optional
+            See :meth:`AbstractHowsoClient.react()`.
+        context_features : Collection of str, optional
+            See :meth:`AbstractHowsoClient.react()`.
+        context_values : Collection of object, optional
+            See :meth:`AbstractHowsoClient.react()`.
+        desired_conviction : float, optional
+            See :meth:`AbstractHowsoClient.react()`.
+        case_indices : Sequence of tuple of {str, int}, optional
+            See :meth:`AbstractHowsoClient.react()`.
+        preserve_feature_values : Collection of str, optional
+            See :meth:`AbstractHowsoClient.react()`.
+        continue_series : bool
+            See :meth:`AbstractHowsoClient.react_series()`.
+
+        Returns
+        -------
+        tuple
+           Updated action_features, actions, context_features, contexts
+        """
+        trainee = self.trainee_cache.get(trainee_id)
+
+        # Validate case_indices if provided
+        if case_indices is not None:
+            util.validate_case_indices(case_indices)
+
+        # validate discriminative-react only parameters
+        if desired_conviction is None:
+            util.validate_list_shape(context_values, 2, "contexts", "list of object")
+            util.validate_list_shape(action_features, 1, "action_features", "str")
+            util.validate_list_shape(context_features, 1, "context_features", "str")
+
+            if context_values is None:
+                # case_indices/preserve_feature_values are not necessary
+                # when using continue_series, as initial_feature/values may be used
+                if not continue_series and (
+                    case_indices is None or preserve_feature_values is None
+                ):
+                    raise ValueError(
+                        "If `contexts` are not specified, both `case_indices` "
+                        "and `preserve_feature_values` must be specified."
+                    )
+
+        # Preprocess context values
+        if context_values is not None and context_features is None:
+            context_features = internals.get_features_from_data(
+                context_values,
+                data_parameter='contexts',
+                features_parameter='context_features')
+        serialized_contexts = serialize_cases(context_values, context_features, trainee.features)
+
+        # Preprocess action values
+        if action_values is not None and action_features is None:
+            util.validate_list_shape(action_values, 2, "actions", "object")
+            action_features = internals.get_features_from_data(
+                action_values,
+                data_parameter='actions',
+                features_parameter='action_features')
+        serialized_actions = serialize_cases(action_values, action_features, trainee.features)
+
+        return action_features, serialized_actions, context_features, serialized_contexts
+
+    def _preprocess_generate_parameters(  # noqa: C901
+        self,
+        trainee_id: str,
+        desired_conviction: float,
+        num_cases_to_generate: int,
+        *,
+        action_features: t.Optional[Collection[str]] = None,
+        case_indices: t.Optional[CaseIndices] = None,
+        context_features: t.Optional[Collection[str]] = None,
+        context_values: t.Optional[TabularData2D] = None,
+    ) -> tuple[Collection[str] | None, TabularData2D | None]:
+        """
+        Helper method to pre-process generative react parameters.
+
+        Updates the passed-in parameters as necessary.
+
+        Parameters
+        ----------
+        trainee_id : str
+            See :meth:`AbstractHowsoClient.react`.
+        desired_conviction : float, optional
+            See :meth:`AbstractHowsoClient.react`.
+        num_cases_to_generate : int, optional
+            See :meth:`AbstractHowsoClient.react`.
+        action_features : Collection of str, optional
+            See :meth:`AbstractHowsoClient.react`.
+        case_indices : Sequence of tuple of {str, int}, optional
+            See :meth:`AbstractHowsoClient.react`.
+        context_features : Collection of str, optional
+            See :meth:`AbstractHowsoClient.react`.
+        context_values : list of object, optional
+            See :meth:`AbstractHowsoClient.react`.
+
+        Returns
+        -------
+        tuple
+            The prepared context_features and context_values.
+        """
+        if case_indices is not None:
+            if len(case_indices) != 1 and \
+                    len(case_indices) != num_cases_to_generate:
+                raise HowsoError(
+                    "The number of `case_indices` provided does not match "
+                    "the number of cases to generate."
+                )
+
+        # Validate case_indices if provided
+        if case_indices is not None:
+            util.validate_case_indices(case_indices)
+
+        util.validate_list_shape(action_features, 1, "action_features", "str")
+
+        if context_values is not None:
+            if len(context_values) != 1 and len(context_values) != num_cases_to_generate:
+                raise HowsoError(
+                    "The number of case `contexts` provided does not match the "
+                    "number of cases to generate."
+                )
+
+        if context_features and not context_values:
+            raise HowsoError(
+                "For generative reacts, when `context_features` are provided, "
+                "`contexts` values must also be provided."
+            )
+
+        if context_features is not None or context_values is not None:
+            if (
+                context_features is None
+                or context_values is None
+                or not isinstance(context_values[0], Sized)
+                or len(context_features) != len(context_values[0])
+            ):
+                raise HowsoError(
+                    "The number of provided context values in `contexts` "
+                    "does not match the number of features in "
+                    "`context_features`."
+                )
+
+        if desired_conviction <= 0:
+            raise HowsoError("Desired conviction must be greater than 0.")
+
+        if self.configuration.verbose:
+            print(f'Generating case from trainee with id: {trainee_id}')
+
+        for i in range(num_cases_to_generate):
+            target_context_values = None
+            if context_values is not None:
+                if len(context_values) == 1:
+                    target_context_values = context_values[0]
+                elif len(context_values) == num_cases_to_generate:
+                    target_context_values = context_values[i]
+
+            if context_features and (
+                not target_context_values or
+                not isinstance(target_context_values, Sized) or
+                len(target_context_values) != len(context_features)
+            ):
+                raise HowsoError(
+                    f"The number of provided context values in `contexts[{i}]` "
+                    "does not match the number of features in `context_features`."
+                )
+
+        return context_features, context_values
+
+    def react_series(  # noqa: C901
+        self,
+        trainee_id: str,
+        *,
+        action_features: t.Optional[Collection[str]] = None,
+        actions: t.Optional[TabularData2D] = None,
+        batch_size: t.Optional[int] = None,
+        case_indices: t.Optional[CaseIndices] = None,
+        contexts: t.Optional[TabularData2D] = None,
+        context_features: t.Optional[Collection[str]] = None,
+        continue_series: bool = False,
+        continue_series_features: t.Optional[Collection[str]] = None,
+        continue_series_values: t.Optional[list[t.Any] | list[list[t.Any]]] = None,
+        derived_action_features: t.Optional[Collection[str]] = None,
+        derived_context_features: t.Optional[Collection[str]] = None,
+        desired_conviction: t.Optional[float] = None,
+        details: t.Optional[Mapping] = None,
+        exclude_novel_nominals_from_uniqueness_check: bool = False,
+        feature_bounds_map: t.Optional[dict] = None,
+        final_time_steps: t.Optional[list[t.Any] | list[list[t.Any]]] = None,
+        generate_new_cases: GenerateNewCases = "no",
+        init_time_steps: t.Optional[list[t.Any] | list[list[t.Any]]] = None,
+        initial_batch_size: t.Optional[int] = None,
+        initial_features: t.Optional[Collection[str]] = None,
+        initial_values: t.Optional[list[t.Any] | list[list[t.Any]]] = None,
+        input_is_substituted: bool = False,
+        leave_case_out: bool = False,
+        max_series_lengths: t.Optional[list[int]] = None,
+        new_case_threshold: NewCaseThreshold = "min",
+        num_series_to_generate: int = 1,
+        ordered_by_specified_features: bool = False,
+        output_new_series_ids: bool = True,
+        preserve_feature_values: t.Optional[Collection[str]] = None,
+        progress_callback: t.Optional[Callable] = None,
+        series_context_features: t.Optional[Collection[str]] = None,
+        series_context_values: t.Optional[list[t.Any] | list[list[t.Any]]] = None,
+        series_id_tracking: SeriesIDTracking = "fixed",
+        series_index: t.Optional[str] = None,
+        series_stop_maps: t.Optional[list[dict[str, dict]]] = None,
+        substitute_output: bool = True,
+        suppress_warning: bool = False,
+        use_case_weights: bool = False,
+        use_regional_model_residuals: bool = True,
+        weight_feature: t.Optional[str] = None
+    ) -> Reaction:
+        """
+        React in a series until a series_stop_map condition is met.
+
+        Aggregates rows of data corresponding to the specified context, action,
+        derived_context and derived_action features, utilizing previous rows to
+        derive values as necessary. Outputs a dict of "action_features" and
+        corresponding "action" where "action" is the completed 'matrix' for the
+        corresponding `action_features` and `derived_action_features`.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to react to.
+        num_series_to_generate : int, default 1
+            The number of series to generate.
+        final_time_steps : list of object, optional
+            The time steps at which to end synthesis. Time-series
+            only. Must provide either one for all series, or exactly one per
+            series.
+        init_time_steps : list of object, optional
+            The time steps at which to begin synthesis. Time-series
+            only. Must provide either one for all series, or exactly one per
+            series.
+        initial_features : iterable of str, optional
+            List of features to condition just the first case in a
+            series, overwrites context_features and derived_context_features
+            for that first case. All specified initial features must be in one
+            of: context_features, action_features, derived_context_features or
+            derived_action_features. If provided a value that isn't in one of
+            those lists, it will be ignored.
+        initial_values : list of list of object, optional
+            2d list of values corresponding to the initial_features,
+            used to condition just the first case in each series. Must provide
+            either one for all series, or exactly one per series.
+        series_stop_maps : list of dict of dict, optional
+            A dictionary of feature name to stop conditions. Must provide either
+            one for all series, or exactly one per series.
+
+            .. TIP::
+                Stop series when value exceeds max or is smaller than min::
+
+                    {"feature_name":  {"min" : 1, "max": 2}}
+
+                Stop series when feature value matches any of the values
+                listed::
+
+                    {"feature_name":  {"values": ["val1", "val2"]}}
+
+        max_series_lengths : list of int, optional
+            maximum size a series is allowed to be.  Default is
+            3 * model_size, a 0 or less is no limit. If forecasting
+            with ``continue_series``, this defines the maximum length of the
+            forecast. Must provide either one for all series, or exactly
+            one per series.
+        continue_series : bool, default False
+            When True will attempt to continue existing series instead of
+            starting new series. If `initial_values` provide series IDs, it
+            will continue those explicitly specified IDs, otherwise it will
+            randomly select series to continue.
+            .. note::
+
+                Terminated series with terminators cannot be continued and
+                will result in null output.
+        continue_series_features : list of str, optional
+            The list of feature names corresponding to the values in each row of
+            `continue_series_values`. This value is ignored if
+            `continue_series_values` is None.
+        continue_series_values : list of list of list of object or list of pandas.DataFrame, default None
+            The set of series data to be forecasted with feature values in the
+            same order defined by `continue_series_values`. The value of
+            `continue_series` will be ignored and treated as true if this value
+            is specified.
+        derived_context_features : iterable of str, optional
+            List of context features whose values should be computed
+            from the entire series in the specified order. Must be
+            different than context_features.
+        derived_action_features : iterable of str, optional
+            List of action features whose values should be computed
+            from the resulting last row in series, in the specified order.
+            Must be a subset of action_features.
+
+            .. note::
+
+                Both of these derived feature lists rely on the features'
+                "derived_feature_code" attribute to compute the values. If
+                "derived_feature_code" attribute references non-existing
+                feature indices, the derived value will be null.
+
+        exclude_novel_nominals_from_uniqueness_check : bool, default False
+            If True, will exclude features which have a subtype defined in their feature
+            attributes from the uniqueness check that happens when ``generate_new_cases``
+            is True. Only applies to generative reacts.
+        series_context_features : iterable of str, optional
+            List of context features corresponding to
+            series_context_values, if specified must not overlap with any
+            initial_features or context_features.
+        series_context_values : list of list of list of object or list of DataFrame, optional
+            3d-list of context values, one for each feature for each
+            row for each series. If specified, max_series_lengths are ignored.
+        output_new_series_ids : bool, default True
+            If True, series ids are replaced with unique values on output.
+            If False, will maintain or replace ids with existing trained values,
+            but also allows output of series with duplicate existing ids.
+        series_id_tracking : {"dynamic", "fixed", "no"}, default "fixed"
+            Controls how closely generated series should follow existing series (plural).
+
+            Choices are: "fixed" , "dynamic" or "no":
+
+            - If "fixed", tracks the particular relevant series ID.
+            - If "dynamic", tracks the particular relevant series ID, but is
+              allowed to change the series ID that it tracks based on its
+              current context.
+            - If "no", does not track any particular series ID.
+        series_index : str, Optional
+            When set to a string, will include the series index as a
+            column in the returned DataFrame using the column name given.
+            If set to None, no column will be added.
+        progress_callback : callable, optional
+            A callback method that will be called before each
+            batched call to react series and at the end of reacting. The method
+            is given a ProgressTimer containing metrics on the progress and
+            timing of the react series operation, and the batch result.
+        batch_size: int, optional
+            Define the number of series to react to at once. If left
+            unspecified, the batch size will be determined automatically.
+        initial_batch_size: int, optional
+            The number of series to react to in the first batch. If unspecified,
+            the number will be determined automatically. The number of series
+            in following batches will be automatically adjusted. This value is
+            ignored if ``batch_size`` is specified.
+        contexts: list of list of object or DataFrame
+            See parameter ``contexts`` in :meth:`HowsoDirectClient.react`.
+        action_features: iterable of str
+            See parameter ``action_features`` in :meth:`HowsoDirectClient.react`.
+        actions: list of list of object or DataFrame
+            See parameter ``actions`` in :meth:`HowsoDirectClient.react`.
+        context_features: iterable of str
+            See parameter ``context_features`` in :meth:`HowsoDirectClient.react`.
+        input_is_substituted : bool, default False
+            See parameter ``input_is_substituted`` in :meth:`HowsoDirectClient.react`.
+        substitute_output : bool
+            See parameter ``substitute_output`` in :meth:`HowsoDirectClient.react`.
+        details: dict, optional
+            See parameter ``details`` in :meth:`HowsoDirectClient.react`.
+        desired_conviction: float
+            See parameter ``desired_conviction`` in :meth:`HowsoDirectClient.react`.
+        weight_feature : str
+            See parameter ``weight_feature`` in :meth:`HowsoDirectClient.react`.
+        use_case_weights : bool
+            See parameter ``use_case_weights`` in :meth:`HowsoDirectClient.react`.
+        case_indices: iterable of sequence of str, int
+            See parameter ``case_indices`` in :meth:`HowsoDirectClient.react`.
+        preserve_feature_values : iterable of str
+            See parameter ``preserve_feature_values`` in :meth:`HowsoDirectClient.react`.
+        new_case_threshold : str
+            See parameter ``new_case_threshold`` in :meth:`HowsoDirectClient.react`.
+        leave_case_out : bool
+            See parameter ``leave_case_out`` in :meth:`HowsoDirectClient.react`.
+        use_regional_model_residuals : bool
+            See parameter ``use_regional_model_residuals`` in :meth:`HowsoDirectClient.react`.
+        feature_bounds_map: dict of dict
+            See parameter ``feature_bounds_map`` in :meth:`HowsoDirectClient.react`.
+        generate_new_cases : {"always", "attempt", "no"}
+            See parameter ``generate_new_cases`` in :meth:`HowsoDirectClient.react`.
+        ordered_by_specified_features : bool
+            See parameter ``ordered_by_specified_features`` in :meth:`HowsoDirectClient.react`.
+        suppress_warning : bool
+            See parameter ``suppress_warning`` in :meth:`HowsoDirectClient.react`.
+
+        Returns
+        -------
+        Reaction:
+            A MutableMapping (dict-like) with these keys -> values:
+                action -> pandas.DataFrame
+                    A data frame of action values.
+
+                details -> Dict or List
+                    An aggregated list of any requested details.
+
+        Raises
+        ------
+        ValueError
+            If the number of provided context values does not match the length of
+            context features.
+
+            If `series_context_values` is not a 3d list of objects.
+
+            If `series_continue_values` is not a 3d list of objects.
+
+            If `derived_action_features` is not a subset of `action_features`.
+
+            If `new_case_threshold` is not one of {"max", "min", "most_similar"}.
+        HowsoError
+            If `num_series_to_generate` is not an integer greater than 0.
+        """
+        trainee_id = self._resolve_trainee(trainee_id)
+        feature_attributes = self.trainee_cache.get(trainee_id).features
+
+        util.validate_list_shape(initial_features, 1, "initial_features", "str")
+        util.validate_list_shape(initial_values, 2, "initial_values", "list of object")
+        util.validate_list_shape(initial_features, 1, "max_series_lengths", "num")
+        util.validate_list_shape(series_stop_maps, 1, "series_stop_maps", "dict")
+        util.validate_list_shape(series_context_features, 1, "series_context_features", "str")
+
+        if continue_series_values and util.num_list_dimensions(continue_series_values) != 3:
+            raise ValueError(
+                "Improper shape of `continue_series_values` values passed. "
+                "`continue_series_values` must be a 3d list of object.")
+        if series_context_values and util.num_list_dimensions(series_context_values) != 3:
+            raise ValueError(
+                "Improper shape of `series_context_values` values passed. "
+                "`series_context_values` must be a 3d list of object.")
+
+        if continue_series_values is not None:
+            continue_series = True
+
+        action_features, actions, context_features, contexts = (
+            self._preprocess_react_parameters(
+                action_features=action_features,
+                action_values=actions,
+                case_indices=case_indices,
+                context_features=context_features,
+                context_values=contexts,
+                desired_conviction=desired_conviction,
+                preserve_feature_values=preserve_feature_values,
+                trainee_id=trainee_id,
+                continue_series=continue_series,
+            )
+        )
+
+        if action_features is not None and derived_action_features is not None:
+            if not set(derived_action_features).issubset(set(action_features)):
+                raise ValueError(
+                    'Specified `derived_action_features` must be a subset of `action_features`.')
+
+        serialized_series_context_values = None
+        if series_context_values:
+            serialized_series_context_values = []
+            for series in series_context_values:
+                if series_context_features is None:
+                    series_context_features = internals.get_features_from_data(
+                        data=series,
+                        data_parameter="series_context_values",
+                        features_parameter="series_context_features")
+                serialized_series_context_values.append(
+                    serialize_cases(series, series_context_features, feature_attributes))
+
+        serialized_continue_series_values = None
+        if continue_series_values:
+            serialized_continue_series_values = []
+            for series in continue_series_values:
+                if continue_series_features is None:
+                    continue_series_features = internals.get_features_from_data(
+                        data=series,
+                        data_parameter="continue_series_values",
+                        features_parameter="continue_series_features")
+                serialized_continue_series_values.append(
+                    serialize_cases(series, continue_series_features, feature_attributes))
+
+        if new_case_threshold not in [None, "min", "max", "most_similar"]:
+            raise ValueError(
+                f"The value '{new_case_threshold}' specified for the parameter "
+                "`new_case_threshold` is not valid. It accepts one of the"
+                " following values - ['min', 'max', 'most_similar',]"
+            )
+
+        if initial_values is not None and initial_features is None:
+            initial_features = internals.get_features_from_data(
+                initial_values,
+                data_parameter='initial_values',
+                features_parameter='initial_features')
+        initial_values = serialize_cases(initial_values, initial_features, feature_attributes)
+
+        # All of these params must be of length 1 or N
+        # where N is the length of the largest
+        one_or_more_params = [
+            contexts,
+            initial_values,
+            serialized_continue_series_values,
+            serialized_series_context_values,
+            case_indices,
+            actions,
+            max_series_lengths,
+            series_stop_maps,
+        ]
+        if any(one_or_more_params):
+            param_lengths = set([len(x) for x in one_or_more_params if x])
+            if len(param_lengths - {1}) > 1:
+                # Raise error if any of the params have different lengths
+                # greater than 1
+                raise ValueError(
+                    'When providing any of `contexts`, `actions`, '
+                    '`series_context_values`, `continue_series_values`, '
+                    '`case_indices`, `initial_values`, `max_series_lengths`, '
+                    'or `series_stop_maps`, each must be of length 1 or the same '
+                    'length as each other.')
+        else:
+            param_lengths = {1}
+
+        if desired_conviction is None:
+            if case_indices and not preserve_feature_values:
+                raise ValueError(
+                    "For discriminative reacts, `preserve_feature_values` "
+                    "is required when `case_indices` is specified.")
+            else:
+                total_size = max(param_lengths)
+
+            react_params = {
+                "action_features": action_features,
+                "action_values": actions,
+                "context_features": context_features,
+                "context_values": contexts,
+                "continue_series": continue_series,
+                "continue_series_features": continue_series_features,
+                "continue_series_values": serialized_continue_series_values,
+                "initial_features": initial_features,
+                "initial_values": initial_values,
+                "final_time_steps": final_time_steps,
+                "init_time_steps": init_time_steps,
+                "series_stop_maps": series_stop_maps,
+                "max_series_lengths": max_series_lengths,
+                "derived_context_features": derived_context_features,
+                "derived_action_features": derived_action_features,
+                "series_context_features": series_context_features,
+                "series_context_values": serialized_series_context_values,
+                "case_indices": case_indices,
+                "preserve_feature_values": preserve_feature_values,
+                "new_case_threshold": new_case_threshold,
+                "input_is_substituted": input_is_substituted,
+                "substitute_output": substitute_output,
+                "weight_feature": weight_feature,
+                "use_case_weights": use_case_weights,
+                "leave_case_out": leave_case_out,
+                "details": details,
+                "exclude_novel_nominals_from_uniqueness_check": exclude_novel_nominals_from_uniqueness_check,
+                "series_id_tracking": series_id_tracking,
+                "output_new_series_ids": output_new_series_ids,
+            }
+
+        else:
+            if (
+                not isinstance(num_series_to_generate, int) or
+                num_series_to_generate <= 0
+            ):
+                raise HowsoError("`num_series_to_generate` must be an integer "
+                                 "greater than 0.")
+            if max(param_lengths) not in [1, num_series_to_generate]:
+                raise ValueError(
+                    'For generative reacts, when specifying parameters with '
+                    'values for each series they must be of length 1 or the '
+                    'value specified by `num_series_to_generate`.')
+            total_size = num_series_to_generate
+
+            context_features, contexts = \
+                self._preprocess_generate_parameters(
+                    trainee_id,
+                    action_features=action_features,
+                    context_features=context_features,
+                    context_values=contexts,
+                    desired_conviction=desired_conviction,
+                    num_cases_to_generate=num_series_to_generate,
+                    case_indices=case_indices
+                )
+
+            react_params = {
+                "num_series_to_generate": num_series_to_generate,
+                "action_features": action_features,
+                "context_features": context_features,
+                "context_values": contexts,
+                "continue_series": continue_series,
+                "continue_series_features": continue_series_features,
+                "continue_series_values": continue_series_values,
+                "initial_features": initial_features,
+                "initial_values": initial_values,
+                "final_time_steps": final_time_steps,
+                "init_time_steps": init_time_steps,
+                "series_stop_maps": series_stop_maps,
+                "max_series_lengths": max_series_lengths,
+                "derived_context_features": derived_context_features,
+                "derived_action_features": derived_action_features,
+                "exclude_novel_nominals_from_uniqueness_check": exclude_novel_nominals_from_uniqueness_check,
+                "series_context_features": series_context_features,
+                "series_context_values": serialized_series_context_values,
+                "use_regional_model_residuals": use_regional_model_residuals,
+                "desired_conviction": desired_conviction,
+                "feature_bounds_map": feature_bounds_map,
+                "generate_new_cases": generate_new_cases,
+                "ordered_by_specified_features": ordered_by_specified_features,
+                "input_is_substituted": input_is_substituted,
+                "substitute_output": substitute_output,
+                "weight_feature": weight_feature,
+                "use_case_weights": use_case_weights,
+                "preserve_feature_values": preserve_feature_values,
+                "new_case_threshold": new_case_threshold,
+                "case_indices": case_indices,
+                "leave_case_out": leave_case_out,
+                "details": details,
+                "series_id_tracking": series_id_tracking,
+                "output_new_series_ids": output_new_series_ids,
+            }
+
+        if batch_size or self._should_react_batch(react_params, total_size):
+            if self.configuration.verbose:
+                print(f'Batch series reacting on trainee with id: {trainee_id}')
+            response = self._batch_react_series(
+                trainee_id,
+                react_params,
+                total_size=total_size,
+                batch_size=batch_size,
+                initial_batch_size=initial_batch_size,
+                progress_callback=progress_callback)
+        else:
+            if self.configuration.verbose:
+                print(f'Series reacting on trainee with id: {trainee_id}')
+            with ProgressTimer(total_size) as progress:
+                if isinstance(progress_callback, Callable):
+                    progress_callback(progress, None)
+                response, _, _ = self._react_series(trainee_id, react_params)
+                progress.update(total_size)
+
+            if isinstance(progress_callback, Callable):
+                progress_callback(progress, response)
+
+        # put all details under the 'details' key
+        action = response.pop('action')
+        response = {'action': action, 'details': response}
+
+        # If the number of series generated is less then requested, raise
+        # warning, for generative reacts
+        if desired_conviction is not None:
+            len_action = len(response['action'])
+            internals.insufficient_generation_check(
+                num_series_to_generate, len_action,
+                suppress_warning=suppress_warning
+            )
+
+        series_df = util.build_react_series_df(response, series_index=series_index)
+        return Reaction(series_df, response.get('details'))
+
+    def _batch_react_series(  # noqa: C901
+        self,
+        trainee_id: str,
+        params: dict,
+        *,
+        batch_size: t.Optional[int] = None,
+        initial_batch_size: t.Optional[int] = None,
+        total_size: int,
+        progress_callback: t.Optional[Callable] = None
+    ):
+        """
+        Make react series requests in batch.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to react to.
+        params : dict
+            The engine react series parameters.
+        batch_size: int, optional
+            Define the number of series to react to at once. If left
+            unspecified, the batch size will be determined automatically.
+        initial_batch_size: int, optional
+            The number of series to react to in the first batch. If unspecified,
+            the number will be determined automatically. The number of series
+            in following batches will be automatically adjusted. This value is
+            ignored if ``batch_size`` is specified.
+        total_size : int
+            The total size of the data that will be batched.
+        progress_callback : callable, optional
+            A function to be called during batching to retrieve or
+            report progress metrics.
+
+        Returns
+        -------
+        dict
+            The `react_series` response.
+        """
+        temp_result = None
+        accumulated_result = {'action_values': []}
+
+        actions = params.get('action_values')
+        contexts = params.get('context_values')
+        case_indices = params.get('case_indices')
+        initial_values = params.get('initial_values')
+        max_series_lengths = params.get('max_series_lengths')
+        series_context_values = params.get('series_context_values')
+        series_stop_maps = params.get('series_stop_maps')
+        continue_values = params.get('continue_series_values')
+
+        with ProgressTimer(total_size) as progress:
+            batch_scaler = None
+            gen_batch_size = None
+            if not batch_size:
+                if not initial_batch_size:
+                    start_batch_size = max(self._get_trainee_thread_count(trainee_id), 1)
+                else:
+                    start_batch_size = initial_batch_size
+                batch_scaler = self.batch_scaler_class(start_batch_size, progress)
+                gen_batch_size = batch_scaler.gen_batch_size()
+                batch_size = next(gen_batch_size, None)
+
+            while not progress.is_complete and batch_size is not None:
+                if isinstance(progress_callback, Callable):
+                    progress_callback(progress, temp_result)
+                batch_start = progress.current_tick
+                batch_end = progress.current_tick + batch_size
+
+                if actions is not None and len(actions) > 1:
+                    params['action_values'] = actions[batch_start:batch_end]
+                if contexts is not None and len(contexts) > 1:
+                    params['context_values'] = contexts[batch_start:batch_end]
+                if case_indices is not None and len(case_indices) > 1:
+                    params['case_indices'] = case_indices[batch_start:batch_end]
+                if initial_values is not None and len(initial_values) > 1:
+                    params['initial_values'] = initial_values[batch_start:batch_end]
+                if max_series_lengths is not None and len(max_series_lengths) > 1:
+                    params['max_series_lengths'] = max_series_lengths[batch_start:batch_end]
+                if series_context_values is not None and len(series_context_values) > 1:
+                    params['series_context_values'] = series_context_values[batch_start:batch_end]
+                if series_stop_maps is not None and len(series_stop_maps) > 1:
+                    params['series_stop_maps'] = series_stop_maps[batch_start:batch_end]
+                if continue_values is not None and len(continue_values) > 1:
+                    params['continue_series_values'] = continue_values[batch_start:batch_end]
+
+                if params.get('desired_conviction') is not None:
+                    params['num_series_to_generate'] = batch_size
+                temp_result, in_size, out_size = self._react_series(trainee_id, params)
+
+                internals.accumulate_react_result(accumulated_result, temp_result)
+                if batch_scaler is None or gen_batch_size is None:
+                    progress.update(batch_size)
+                else:
+                    batch_size = batch_scaler.send(
+                        gen_batch_size,
+                        batch_scaler.SendOptions(None, (in_size, out_size)))
+
+        # Final call to callback on completion
+        if isinstance(progress_callback, Callable):
+            progress_callback(progress, temp_result)
+
+        return accumulated_result
+
+    def _react_series(self, trainee_id: str, params: dict):
+        """
+        Make a single react series request.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The id of the trainee.
+        params : dict
+            The engine react series parameters.
+
+        Returns
+        -------
+        dict
+            The react series response.
+        int
+            The request payload size.
+        int
+            The response payload size.
+        """
+        batch_result, in_size, out_size = self.execute(trainee_id, "react_series", params)
+
+        if batch_result is None or batch_result.get('action_values') is None:
+            raise ValueError('Invalid parameters passed to react_series.')
+
+        ret = dict()
+        batch_result = util.replace_doublemax_with_infinity(batch_result)
+
+        # batch_result always has action_features and action_values
+        ret['action_features'] = batch_result.pop('action_features') or []
+        ret['action'] = batch_result.pop('action_values')
+
+        # ensure all the details items are output as well
+        for k, v in batch_result.items():
+            ret[k] = [] if v is None else v
+
+        return ret, in_size, out_size
+
     def react_into_features(
-        self, trainee_id, *,
+        self,
+        trainee_id: str,
+        *,
         distance_contribution: bool | str = False,
         familiarity_conviction_addition: bool | str = False,
         familiarity_conviction_removal: bool | str = False,
-        features=None,
+        features: t.Optional[Collection[str]] = None,
         influence_weight_entropy: bool | str = False,
         p_value_of_addition: bool | str = False,
         p_value_of_removal: bool | str = False,
         similarity_conviction: bool | str = False,
-        use_case_weights: bool | str = False,
-        weight_feature=None
+        use_case_weights: bool = False,
+        weight_feature: t.Optional[str] = None,
     ):
-        """Calculate conviction and other data for the specified feature(s)."""
+        """
+        Calculate and cache conviction and other statistics.
 
-    @abstractmethod
+        Parameters
+        ----------
+        trainee_id : str
+            The ID of the Trainee to calculate and store conviction for.
+        features : iterable of str, optional
+            An iterable of features to calculate convictions.
+        familiarity_conviction_addition : bool or str, default False
+            The name of the feature to store conviction of addition
+            values. If set to True the values will be stored to the feature
+            'familiarity_conviction_addition'.
+        familiarity_conviction_removal : bool or str, default False
+            The name of the feature to store conviction of removal
+            values. If set to True the values will be stored to the feature
+            'familiarity_conviction_removal'.
+        influence_weight_entropy : bool or str, default False
+            The name of the feature to store influence weight entropy values in.
+            If set to True, the values will be stored in the feature
+            'influence_weight_entropy'.
+        p_value_of_addition : bool or str, default False
+            The name of the feature to store p value of addition
+            values. If set to True the values will be stored to the feature
+            'p_value_of_addition'.
+        p_value_of_removal : bool or str, default False
+            The name of the feature to store p value of removal
+            values. If set to True the values will be stored to the feature
+            'p_value_of_removal'.
+        similarity_conviction : bool or str, default False
+            The name of the feature to store similarity conviction
+            values. If set to True the values will be stored to the feature
+            'similarity_conviction'.
+        distance_contribution : bool or str, default False
+            The name of the feature to store distance contribution.
+            If set to True the values will be stored to the
+            feature 'distance_contribution'.
+        weight_feature : str, optional
+            Name of feature whose values to use as case weights.
+            When left unspecified uses the internally managed case weight.
+        use_case_weights : bool, default False
+            If set to True will scale influence weights by each
+            case's weight_feature weight.
+        """
+        trainee_id = self._resolve_trainee(trainee_id)
+        util.validate_list_shape(features, 1, "features", "str")
+        if self.configuration.verbose:
+            print(f'Reacting into features on Trainee with id: {trainee_id}')
+        self.execute(trainee_id, "react_into_features", {
+            "features": features,
+            "familiarity_conviction_addition": familiarity_conviction_addition,
+            "familiarity_conviction_removal": familiarity_conviction_removal,
+            "influence_weight_entropy": influence_weight_entropy,
+            "p_value_of_addition": p_value_of_addition,
+            "p_value_of_removal": p_value_of_removal,
+            "similarity_conviction": similarity_conviction,
+            "distance_contribution": distance_contribution,
+            "weight_feature": weight_feature,
+            "use_case_weights": use_case_weights,
+        })
+        self._auto_persist_trainee(trainee_id)
+
     def react_aggregate(
-        self, trainee_id, *,
-        action_feature=None,
-        confusion_matrix_min_count=None,
-        context_features=None,
-        details=None,
-        feature_influences_action_feature=None,
-        hyperparameter_param_path=None,
-        num_samples=None,
-        num_robust_influence_samples=None,
-        num_robust_residual_samples=None,
-        num_robust_influence_samples_per_case=None,
-        prediction_stats_action_feature=None,
-        residuals_hyperparameter_feature=None,
-        robust_hyperparameters=None,
-        sample_model_fraction=None,
-        sub_model_size=None,
-        use_case_weights=None,
-        weight_feature=None,
-    ) -> DataFrame | dict:
-        """Computes, caches, and/or returns specified feature interpretations."""
+        self,
+        trainee_id: str,
+        *,
+        action_feature: t.Optional[str] = None,
+        confusion_matrix_min_count: t.Optional[int] = None,
+        context_features: t.Optional[Collection[str]] = None,
+        details: t.Optional[dict] = None,
+        feature_influences_action_feature: t.Optional[str] = None,
+        hyperparameter_param_path: t.Optional[Collection[str]] = None,
+        num_robust_influence_samples: t.Optional[int] = None,
+        num_robust_residual_samples: t.Optional[int] = None,
+        num_robust_influence_samples_per_case: t.Optional[int] = None,
+        num_samples: t.Optional[int] = None,
+        prediction_stats_action_feature: t.Optional[str] = None,
+        residuals_hyperparameter_feature: t.Optional[str] = None,
+        robust_hyperparameters: t.Optional[bool] = None,
+        sample_model_fraction: t.Optional[float] = None,
+        sub_model_size: t.Optional[int] = None,
+        use_case_weights: bool = False,
+        weight_feature: t.Optional[str] = None,
+    ) -> dict[str, dict[str, float]]:
+        """
+        Reacts into the aggregate trained cases in the Trainee.
 
-    @abstractmethod
+        Calculates, caches, and/or returns the requested influences and prediction stats.
+
+        Parameters
+        ----------
+        action_feature : str, optional
+            Name of target feature for which to do computations. If ``prediction_stats_action_feature``
+            and ``feature_influences_action_feature`` are not provided, they will default to this value.
+            If ``feature_influences_action_feature`` is not provided and feature influences ``details`` are
+            selected, this feature must be provided.
+        confusion_matrix_min_count : int, optional
+            The number of predictions a class should have (value of a cell in the
+            matrix) for it to remain in the confusion matrix. If the count is
+            less than this value, it will be accumulated into a single value of
+            all insignificant predictions for the class and removed from the
+            confusion matrix. Defaults to 10, applicable only to confusion
+            matrices when computing residuals.
+        context_features : iterable of str, optional
+            List of features names to use as contexts for
+            computations. Default is all trained non-unique features if
+            unspecified.
+        details : map of str -> object, optional
+            If details are specified, the response will contain the requested
+            explanation data.. Below are the valid keys and data types for the
+            different audit details. Omitted keys, values set to None, or False
+            values for Booleans will not be included in the data returned.
+
+            - prediction_stats : bool, optional
+                If True outputs full feature prediction stats for all (context and action) features.
+                The prediction stats returned are set by the "selected_prediction_stats" parameter
+                in the `details` parameter. Uses full calculations, which uses leave-one-out for
+                features for computations. False removes cached values.
+            - feature_residuals_full : bool, optional
+                For each context_feature, use the full set of all other context_features to predict
+                the feature. False removes cached values. When ``prediction_stats``
+                in the ``details`` parameter is true, the Trainee will also calculate and cache the
+                full feature residuals.
+            - feature_residuals_robust : bool, optional
+                For each context_feature, use the robust (power set/permutations) set of all other
+                context_features to predict the feature. False removes cached values.
+            - feature_contributions_full : bool, optional
+                For each context_feature, use the full set of all other
+                context_features to compute the mean absolute delta between
+                prediction of action feature with and without the context features
+                in the model. False removes cached values.
+            - feature_contributions_robust : bool, optional
+                For each context_feature, use the robust (power set/permutation)
+                set of all other context_features to compute the mean absolute
+                delta between prediction of the action feature with and without the
+                context features in the model. False removes cached values.
+            - feature_mda_full : bool, optional
+                When True will compute Mean Decrease in Accuracy (MDA)
+                for each context feature at predicting the action feature. Drop
+                each feature and use the full set of remaining context features
+                for each prediction. False removes cached values.
+            - feature_mda_robust : bool, optional
+                Compute Mean Decrease in Accuracy MDA by dropping each feature and using the
+                robust (power set/permutations) set of remaining context features
+                for each prediction. False removes cached values.
+            - feature_feature_mda_permutation_full : bool, optional
+                Compute MDA by scrambling each feature and using the
+                full set of remaining context features for each prediction.
+                False removes cached values.
+            - feature_feature_mda_permutation_robust : bool, optional
+                Compute MDA by scrambling each feature and using the
+                robust (power set/permutations) set of remaining context features
+                for each prediction. False removes cached values.
+            - action_condition : map of str -> any, optional
+                A condition map to select the action set, which is the dataset for which
+                the prediction stats are for. If both ``action_condition`` and ``context_condition``
+                are provided, then all of the action cases selected by the ``action_condition``
+                will be excluded from the context set, which is the set being queried to make to
+                make predictions on the action set, effectively holding them out.
+                If only ``action_condition`` is specified, then only the single predicted case
+                will be left out.
+
+                .. NOTE::
+                    The dictionary keys are the feature name and values are one of:
+
+                        - None
+                        - A value, must match exactly.
+                        - An array of two numeric values, specifying an inclusive
+                        range. Only applicable to continuous and numeric ordinal
+                        features.
+                        - An array of string values, must match any of these values
+                        exactly. Only applicable to nominal and string ordinal
+                        features.
+            - action_num_cases : int, optional
+                The maximum amount of cases to use to calculate prediction stats.
+                If not specified, the limit will be k cases if precision is
+                "similar", or 1000 cases if precision is "exact". Works with or
+                without ``action_condition``.
+                -If ``action_condition`` is set:
+                    If None, will be set to k if precision is "similar" or no limit if precision is "exact".
+                - If ``action_condition`` is not set:
+                    If None, will be set to the Howso default limit of 2000.
+            - action_condition_precision : {"exact", "similar"}, optional
+                The precision to use when selecting cases with the ``action_condition``.
+                If not specified "exact" will be used. Only used if ``action_condition``
+                is not None.
+            - context_condition : map of str -> any, optional
+                A condition map to select the context set, which is the set being queried to make
+                to make predictions on the action set. If both ``action_condition`` and ``context_condition``
+                are provided,  then all of the cases from the action set, which is the dataset for which the
+                prediction stats are for, will be excluded from the context set, effectively holding them out.
+                If only ``action_condition`` is specified,  then only the single predicted case will be left out.
+
+                .. NOTE::
+                    The dictionary keys are the feature name and values are one of:
+
+                        - None
+                        - A value, must match exactly.
+                        - An array of two numeric values, specifying an inclusive
+                        range. Only applicable to continuous and numeric ordinal
+                        features.
+                        - An array of string values, must match any of these values
+                        exactly. Only applicable to nominal and string ordinal
+                        features.
+            - context_precision_num_cases : int, optional
+                Limit on the number of context cases when ``context_condition_precision`` is set to "similar".
+                If None, will be set to k.
+            - context_condition_precision : {"exact", "similar"}, optional
+                The precision to use when selecting cases with the ``context_condition``.
+                If not specified "exact" will be used. Only used if ``context_condition``
+                is not None.
+            - prediction_stats_features : list, optional
+                List of features to use when calculating conditional prediction stats. Should contain all
+                action and context features desired. If ``action_feature`` is also provided, that feature will
+                automatically be appended to this list if it is not already in the list.
+                    stats : list of str, optional
+            - missing_value_accuracy_full : bool, optional
+                The number of cases with missing values predicted to have missing values divided by the number
+                of cases with missing values, applies to all features that contain missing values. Uses full
+                calculations.
+            - missing_value_accuracy_robust : bool, optional
+                The number of cases with missing values predicted to have missing values divided by the number
+                of cases with missing values, applies to all features that contain missing values. Uses robust
+                calculations.
+            - selected_prediction_stats : list, optional
+                List of stats to output. When unspecified, returns all except the confusion matrix. Allowed values:
+
+                - all : Returns all the the available prediction stats, including the confusion matrix.
+                - accuracy : The number of correct predictions divided by the
+                  total number of predictions.
+                - confusion_matrix : A sparse map of actual feature value to a map of
+                  predicted feature value to counts.
+                - mae : Mean absolute error. For continuous features, this is
+                  calculated as the mean of absolute values of the difference
+                  between the actual and predicted values. For nominal features,
+                  this is 1 - the average categorical action probability of each case's
+                  correct classes. Categorical action probabilities are the probabilities
+                  for each class for the action feature.
+                - mda : Mean decrease in accuracy when each feature is dropped
+                  from the model, applies to all features.
+                - feature_mda_permutation_full : Mean decrease in accuracy that used
+                  scrambling of feature values instead of dropping each
+                  feature, applies to all features.
+                - precision : Precision (positive predictive) value for nominal
+                  features only.
+                - r2 : The r-squared coefficient of determination, for
+                  continuous features only.
+                - recall : Recall (sensitivity) value for nominal features only.
+                - rmse : Root mean squared error, for continuous features only.
+                - spearman_coeff : Spearman's rank correlation coefficient,
+                  for continuous features only.
+                - mcc : Matthews correlation coefficient, for nominal features only.
+        feature_influences_action_feature : str, optional
+            When feature influences such as contributions and mda, use this feature as
+            the action feature.  If not provided, will default to the ``action_feature`` if provided.
+            If ``action_feature`` is not provided and feature influences ``details`` are
+            selected, this feature must be provided.
+        hyperparameter_param_path : iterable of str, optional.
+            Full path for hyperparameters to use for computation. If specified
+            for any residual computations, takes precedence over action_feature
+            parameter.  Can be set to a 'paramPath' value from the results of
+            'get_params()' for a specific set of hyperparameters.
+        num_robust_influence_samples : int, optional
+            Total sample size of model to use (using sampling with replacement)
+            for robust contribution computation. Defaults to 300.
+        num_robust_residual_samples : int, optional
+            Total sample size of model to use (using sampling with replacement)
+            for robust mda and residual computation.
+            Defaults to 1000 * (1 + log(number of features)).  Note: robust mda
+            will be updated to use num_robust_influence_samples in a future release.
+        num_robust_influence_samples_per_case : int, optional
+            Specifies the number of robust samples to use for each case for
+            robust contribution computations.
+            Defaults to 300 + 2 * (number of features).
+        num_samples : int, optional
+            Total sample size of model to use (using sampling with replacement)
+            for all non-robust computation. Defaults to 1000.
+            If specified overrides sample_model_fraction.```
+        residuals_hyperparameter_feature : str, optional
+            When calculating residuals and prediction stats, uses this target
+            features's hyperparameters. The trainee must have been analyzed with
+            this feature as the action feature first. If not provided, by default
+            residuals and prediction stats uses ".targetless" hyperparameters.
+        robust_hyperparameters : bool, optional
+            When specified, will attempt to return residuals that were
+            computed using hyperparameters with the specified robust or
+            non-robust type.
+        prediction_stats_action_feature : str, optional
+            When calculating residuals and prediction stats, uses this target features's
+            hyperparameters. The trainee must have been analyzed with this feature as the
+            action feature first. If both ``prediction_stats_action_feature`` and
+            ``action_feature`` are not provided, by default residuals and prediction
+            stats uses ".targetless" hyperparameters. If "action_feature" is provided,
+            and this value is not provided, will default to ``action_feature``.
+        sample_model_fraction : float, optional
+            A value between 0.0 - 1.0, percent of model to use in sampling
+            (using sampling without replacement). Applicable only to non-robust
+            computation. Ignored if num_samples is specified.
+            Higher values provide better accuracy at the cost of compute time.
+        sub_model_size : int, optional
+            Subset of model to use for calculations. Applicable only
+            to models > 1000 cases.
+        use_case_weights : bool, default False
+            If set to True will scale influence weights by each case's
+            weight_feature weight.
+        weight_feature : str, optional
+            The name of feature whose values to use as case weights.
+            When left unspecified uses the internally managed case weight.
+
+        Returns
+        -------
+        dict of str to dict of str to float
+            If specified, a map of feature to map of stat type to stat values is returned.
+        """
+        trainee_id = self._resolve_trainee(trainee_id)
+        util.validate_list_shape(context_features, 1, "context_features", "str")
+
+        if isinstance(details, dict):
+            if action_condition_precision := details.get("action_condition_precision"):
+                if action_condition_precision not in self.SUPPORTED_PRECISION_VALUES:
+                    warnings.warn(self.WARNING_MESSAGES['invalid_precision'].format("action_condition_precision"))
+
+            if context_condition_precision := details.get("context_condition_precision"):
+                if context_condition_precision not in self.SUPPORTED_PRECISION_VALUES:
+                    warnings.warn(self.WARNING_MESSAGES['invalid_precision'].format("context_condition_precision"))
+
+        if self.configuration.verbose:
+            print(f'Reacting into aggregate trained cases of Trainee with id: {trainee_id}')
+
+        stats = self.execute(trainee_id, "react_aggregate", {
+            "action_feature": action_feature,
+            "residuals_hyperparameter_feature": residuals_hyperparameter_feature,
+            "context_features": context_features,
+            "confusion_matrix_min_count": confusion_matrix_min_count,
+            "details": details,
+            "feature_influences_action_feature": feature_influences_action_feature,
+            "hyperparameter_param_path": hyperparameter_param_path,
+            "num_robust_influence_samples": num_robust_influence_samples,
+            "num_robust_residual_samples": num_robust_residual_samples,
+            "num_robust_influence_samples_per_case": num_robust_influence_samples_per_case,
+            "num_samples": num_samples,
+            "prediction_stats_action_feature": prediction_stats_action_feature,
+            "robust_hyperparameters": robust_hyperparameters,
+            "sample_model_fraction": sample_model_fraction,
+            "sub_model_size": sub_model_size,
+            "use_case_weights": use_case_weights,
+            "weight_feature": weight_feature,
+        })
+        if stats is None:
+            stats = dict()
+
+        self._auto_persist_trainee(trainee_id)
+        return stats
+
     def react_group(
-        self, trainee_id, new_cases, *,
-        distance_contributions=False,
-        familiarity_conviction_addition=True,
-        familiarity_conviction_removal=False,
-        features=None,
-        kl_divergence_addition=False,
-        kl_divergence_removal=False,
-        p_value_of_addition=False,
-        p_value_of_removal=False,
-        use_case_weights=False,
-        weight_feature=None
-    ) -> DataFrame | dict:
-        """Compute specified data for a **set** of cases."""
+        self,
+        trainee_id: str,
+        new_cases: TabularData3D,
+        *,
+        features: t.Optional[Collection[str]] = None,
+        distance_contributions: bool = False,
+        familiarity_conviction_addition: bool = True,
+        familiarity_conviction_removal: bool = False,
+        kl_divergence_addition: bool = False,
+        kl_divergence_removal: bool = False,
+        p_value_of_addition: bool = False,
+        p_value_of_removal: bool = False,
+        weight_feature: t.Optional[str] = None,
+        use_case_weights: bool = False
+    ) -> dict:
+        """
+        Computes specified data for a **set** of cases.
 
-    @abstractmethod
-    def react(
-        self, trainee_id, *,
-        action_features=None,
-        actions=None,
-        allow_nulls=False,
-        batch_size=None,
-        case_indices=None,
-        contexts=None,
-        context_features=None,
-        derived_action_features=None,
-        derived_context_features=None,
-        desired_conviction=None,
-        details=None,
-        exclude_novel_nominals_from_uniqueness_check=False,
-        feature_bounds_map=None,
-        generate_new_cases="no",
-        initial_batch_size=None,
-        input_is_substituted=False,
-        into_series_store=None,
-        leave_case_out=None,
-        new_case_threshold="min",
-        num_cases_to_generate=1,
-        ordered_by_specified_features=False,
-        post_process_features=None,
-        post_process_values=None,
-        preserve_feature_values=None,
-        progress_callback=None,
-        substitute_output=True,
-        suppress_warning=False,
-        use_case_weights=False,
-        use_regional_model_residuals=True,
-        weight_feature=None
-    ) -> Reaction:
-        """Send a `react` to the Howso engine."""
+        Return the list of familiarity convictions (and optionally, distance
+        contributions or p values) for each set.
+
+        Parameters
+        ----------
+        trainee_id : str
+            The trainee id.
+
+        new_cases : list of list of list of object or list of DataFrame
+            Specify a **set** using a list of cases to compute the conviction of
+            groups of cases as shown in the following example.
+
+            >>> [ [[1, 2, 3], [4, 5, 6], [7, 8, 9]], # Group 1
+            >>>   [[1, 2, 3]] ] # Group 2
+
+        features : Collection of str, optional
+            The feature names to consider while calculating convictions.
+        distance_contributions : bool, default False
+            Calculate and output distance contribution ratios in
+            the output dict for each case.
+        familiarity_conviction_addition : bool, default True
+            Calculate and output familiarity conviction of adding the
+            specified cases.
+        familiarity_conviction_removal : bool, default False
+            Calculate and output familiarity conviction of removing
+            the specified cases.
+        kl_divergence_addition : bool, default False
+            Calculate and output KL divergence of adding the
+            specified cases.
+        kl_divergence_removal : bool, default False
+            Calculate and output KL divergence of removing the
+            specified cases.
+        p_value_of_addition : bool, default False
+            If true will output p value of addition.
+        p_value_of_removal : bool, default False
+            If true will output p value of removal.
+        weight_feature : str, optional
+            Name of feature whose values to use as case weights.
+            When left unspecified uses the internally managed case weight.
+        use_case_weights : bool, default False
+            If set to True will scale influence weights by each
+            case's weight_feature weight.
+
+        Returns
+        -------
+        dict
+            The react group response.
+        """
+        trainee_id = self._resolve_trainee(trainee_id)
+        feature_attributes = self.trainee_cache.get(trainee_id).features
+        serialized_cases = None
+
+        if util.num_list_dimensions(new_cases) != 3:
+            raise ValueError(
+                "Improper shape of `new_cases` values passed. "
+                "`new_cases` must be a 3d list of object.")
+
+        serialized_cases = []
+        for group in new_cases:
+            if features is None:
+                features = internals.get_features_from_data(group)
+            serialized_cases.append(serialize_cases(group, features, feature_attributes))
+
+        if self.configuration.verbose:
+            print(f'Reacting to a set of cases on Trainee with id: {trainee_id}')
+        result = self.execute(trainee_id, "react_group", {
+            "features": features,
+            "new_cases": serialized_cases,
+            "distance_contributions": distance_contributions,
+            "familiarity_conviction_addition": familiarity_conviction_addition,
+            "familiarity_conviction_removal": familiarity_conviction_removal,
+            "kl_divergence_addition": kl_divergence_addition,
+            "kl_divergence_removal": kl_divergence_removal,
+            "p_value_of_addition": p_value_of_addition,
+            "p_value_of_removal": p_value_of_removal,
+            "weight_feature": weight_feature,
+            "use_case_weights": use_case_weights,
+        })
+        if result is None:
+            result = dict()
+        return result
 
     def evaluate(
         self,
@@ -1940,7 +4007,7 @@ class AbstractHowsoClient(ABC):
             util.validate_case_indices(case_indices)
 
         if precision is not None and precision not in self.SUPPORTED_PRECISION_VALUES:
-            warnings.warn(self.WARNING_MESSAGES['invalid_precision'])
+            warnings.warn(self.WARNING_MESSAGES['invalid_precision'].format("precision"))
 
         util.validate_list_shape(features, 1, "features", "str")
         if session is None and case_indices is None:
