+++ howso/direct/core.py
@@ -305,28 +305,6 @@ class HowsoCore:
             raise HowsoError("Error loading the Trainee.")
         return {"name": trainee_id}
 
-    def get_loaded_trainees(self) -> List[str]:
-        """
-        Get loaded Trainees.
-
-        Returns
-        -------
-        list of str
-            A list of trainee identifiers that are currently loaded.
-        """
-        return self.get_entities()
-
-    def get_entities(self) -> List[str]:
-        """
-        Get loaded entities.
-
-        Returns
-        -------
-        list of str
-            A list of entity identifiers that are currently loaded.
-        """
-        return self.amlg.get_entities()
-
     def load(
         self,
         trainee_id: str,
@@ -592,23 +570,6 @@ class HowsoCore:
             "trainee_id": subtrainee_id
         })
 
-    def remove_series_store(self, trainee_id: str, series: Optional[str] = None
-                            ) -> None:
-        """
-        Delete part or all of the series store from a Trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee to delete the series store from.
-        series : str, optional
-            The ID of the series to remove from the series store.
-            If None, the entire series store will be deleted.
-        """
-        return self._execute(trainee_id, "remove_series_store", {
-            "series": series,
-        })
-
     def clean_data(
         self,
         trainee_id: str,
@@ -724,51 +685,6 @@ class HowsoCore:
             }
         )
 
-    def compute_feature_weights(
-        self,
-        trainee_id: str,
-        action_feature: Optional[str] = None,
-        context_features: Optional[Iterable[str]] = None,
-        robust: bool = False,
-        weight_feature: Optional[str] = None,
-        use_case_weights: bool = False
-    ) -> Dict:
-        """
-        Compute feature weights for specified context and action features.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        action_feature : str, optional
-            Action feature for which to set the specified feature weights for.
-        context_features: iterable of str
-            List of context feature names.
-        robust : bool, default False.
-            When true, the power set/permutations of features are used as
-            contexts to calculate the residual for a given feature. When
-            false, the full set of features is used to calculate the
-            residual for a given feature.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each
-            case's weight_feature weight.
-
-        Returns
-        -------
-        dict
-            A dictionary of computed context features -> weights.
-        """
-        return self._execute(trainee_id, "compute_feature_weights", {
-            "action_feature": action_feature,
-            "context_features": context_features,
-            "robust": robust,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-        })
-
     def set_metadata(self, trainee_id: str, metadata: Union[Dict, None]
                      ) -> None:
         """
@@ -784,22 +700,6 @@ class HowsoCore:
         return self._execute(trainee_id, "set_metadata", {
             "metadata": metadata,
         })
-
-    def get_metadata(self, trainee_id: str) -> Union[Dict, None]:
-        """
-        Get trainee metadata.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-
-        Returns
-        -------
-        dict or None
-            The metadata dictionary.
-        """
-        return self._execute(trainee_id, "get_metadata", {})
 
     def clear_imputed_session(
         self,
@@ -825,1000 +725,6 @@ class HowsoCore:
         return self._execute(trainee_id, "clear_imputed_session", {
             "session": session,
             "impute_session": impute_session,
-        })
-
-    def append_to_series_store(
-        self,
-        trainee_id: str,
-        series: str,
-        contexts: List[List[Any]],
-        *,
-        context_features: Optional[Iterable[str]] = None
-    ) -> None:
-        """
-        Append the specified contexts to a series store.
-
-        For use with train series.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee to append to.
-        series : str
-            The name of the series store to append to.
-        contexts : list of list of object
-            The list of list of context values to append to the series.
-        context_features : iterable of str, optional
-            The list of feature names for contexts.
-        """
-        return self._execute(
-            trainee_id, "append_to_series_store",
-            {
-                "context_features": context_features,
-                "context_values": contexts,
-                "series": series,
-            }
-        )
-
-    def react(
-        self,
-        trainee_id: str,
-        *,
-        action_features: Optional[Iterable[str]] = None,
-        action_values: Optional[List[List[object]]] = None,
-        allow_nulls: bool = False,
-        case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        context_features: Optional[Iterable[str]] = None,
-        context_values: Optional[List[List[object]]] = None,
-        derived_action_features: Optional[Iterable[str]] = None,
-        derived_context_features: Optional[Iterable[str]] = None,
-        desired_conviction: Optional[float] = None,
-        details: Optional[Dict] = None,
-        exclude_novel_nominals_from_uniqueness_check: bool = False,
-        extra_features: Optional[Iterable[str]] = None,
-        feature_bounds_map: Optional[Dict] = None,
-        generate_new_cases: Literal["always", "attempt", "no"] = "no",
-        input_is_substituted: bool = False,
-        into_series_store: Optional[str] = None,
-        leave_case_out: bool = False,
-        new_case_threshold: Literal["max", "min", "most_similar"] = "min",
-        ordered_by_specified_features: bool = False,
-        post_process_features: Optional[Iterable[str]] = None,
-        post_process_values: Optional[List[object]] = None,
-        preserve_feature_values: Optional[Iterable[str]] = None,
-        substitute_output: bool = True,
-        use_case_weights: bool = False,
-        use_regional_model_residuals: bool = True,
-        weight_feature: Optional[str] = None
-    ) -> Dict:
-        """
-        Single case react.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        context_values : list of list of object, optional
-            A 2d list of context values to react to.
-            If None for discriminative react, it is assumed that `session`
-            and `session_id` keys are set in the `details`.
-        action_features : iterable of str, optional
-            An iterable of feature names to treat as action features during
-            react.
-        action_values : list of list of object, optional
-            One or more action values to use for action features.
-            If specified, will only return the specified explanation
-            details for the given actions. (Discriminative reacts only)
-        allow_nulls : bool, default False
-            When true will allow return of null values if there
-            are nulls in the local model for the action features, applicable
-            only to discriminative reacts.
-        context_features : iterable of str, optional
-            An iterable of feature names to treat as context features during
-            react.
-        derived_context_features : iterable of str, optional
-            An iterable of feature names whose values should be computed
-            from the provided context in the specified order. Must be different
-            than context_features.
-        derived_action_features : iterable of str, optional
-            An iterable of feature names whose values should be computed
-            after generation from the generated case prior to output, in the
-            specified order. Must be a subset of action_features.
-        input_is_substituted : bool, default False
-            if True assumes provided categorical (nominal or
-            ordinal) feature values have already been substituted.
-        substitute_output : bool, default True
-            If False, will not substitute categorical feature
-            values. Only applicable if a substitution value map has been set.
-        details : dict, optional
-            If details are specified, the response will
-            contain the requested explanation data along with the reaction.
-        desired_conviction : float
-            If specified will execute a generative react. If not
-            specified will executed a discriminative react. Conviction is the
-            ratio of expected surprisal to generated surprisal for each
-            feature generated, valid values are in the range of zero to
-            infinity.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each
-            case's weight_feature weight.
-        case_indices : Iterable of Sequence[Union[str, int]], defaults to None
-            An Iterable of Sequences, of session id and index, where
-            index is the original 0-based index of the case as it was trained
-            into the session. If this case does not exist, discriminative react
-            outputs null, generative react ignores it.
-        post_process_features : iterable of str, optional
-            List of feature names that will be made available during the
-            execution of post_process feature attributes.
-        post_process_values : list of object, optional
-            A 2d list of values corresponding to post_process_features that
-            will be made available during the execution of post_process feature
-            attributes.
-        preserve_feature_values : iterable of str
-            List of features that will preserve their values from
-            the case specified by case_indices, appending and overwriting the
-            specified contexts as necessary.  For generative reacts, if
-            case_indices isn't specified will preserve feature values of a
-            random case.
-        leave_case_out : bool, default False
-            If set to True and specified along with case_indices,
-            each individual react will respectively ignore the corresponding
-            case specified by case_indices by leaving it out.
-        into_series_store : str, optional
-            The name of a series store. If specified, will store an internal
-            record of all react contexts for this session and series to be used
-            later with train series.
-        use_regional_model_residuals : bool
-            If false uses model feature residuals, if True
-            recalculates regional model residuals.
-        feature_bounds_map : dict of dict
-            A mapping of feature names to the bounds for the
-            feature values to be generated in.
-        generate_new_cases : {"always", "attempt", "no"}, default "no"
-            How to generate new cases.
-        ordered_by_specified_features : bool, default False
-            If True order of generated feature values will match
-            the order of specified features.
-        new_case_threshold : {"min", "max", "most_similar"}, optional
-            Distance to determine the privacy cutoff. If None,
-            will default to "min".
-        exclude_novel_nominals_from_uniqueness_check : bool, default False
-            If True, will exclude features which have a subtype defined in their feature
-            attributes from the uniqueness check that happens when ``generate_new_cases``
-            is True. Only applies to generative reacts.
-
-        Returns
-        -------
-        dict
-            The react result including audit details.
-        """
-        return self._execute(trainee_id, "react", {
-            "context_features": context_features,
-            "context_values": context_values,
-            "action_features": action_features,
-            "action_values": action_values,
-            "details": details,
-            "derived_action_features": derived_action_features,
-            "derived_context_features": derived_context_features,
-            "exclude_novel_nominals_from_uniqueness_check": exclude_novel_nominals_from_uniqueness_check,
-            "extra_features": extra_features,
-            "case_indices": case_indices,
-            "allow_nulls": allow_nulls,
-            "input_is_substituted": input_is_substituted,
-            "substitute_output": substitute_output,
-            "weight_feature": weight_feature,
-            "leave_case_out": leave_case_out,
-            "use_case_weights": use_case_weights,
-            "use_regional_model_residuals": use_regional_model_residuals,
-            "desired_conviction": desired_conviction,
-            "feature_bounds_map": feature_bounds_map,
-            "generate_new_cases": generate_new_cases,
-            "ordered_by_specified_features": ordered_by_specified_features,
-            "post_process_features": post_process_features,
-            "post_process_values": post_process_values,
-            "preserve_feature_values": preserve_feature_values,
-            "new_case_threshold": new_case_threshold,
-            "into_series_store": into_series_store,
-        })
-
-    def batch_react(
-        self,
-        trainee_id: str,
-        *,
-        action_features: Optional[Iterable[str]] = None,
-        action_values: Optional[List[List[object]]] = None,
-        allow_nulls: bool = False,
-        case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        context_features: Optional[Iterable[str]] = None,
-        context_values: Optional[List[List[object]]] = None,
-        derived_action_features: Optional[Iterable[str]] = None,
-        derived_context_features: Optional[Iterable[str]] = None,
-        desired_conviction: Optional[float] = None,
-        details: Optional[Dict] = None,
-        exclude_novel_nominals_from_uniqueness_check: bool = False,
-        extra_features: Optional[Iterable[str]] = None,
-        feature_bounds_map: Optional[Dict] = None,
-        generate_new_cases: Literal["always", "attempt", "no"] = "no",
-        input_is_substituted: bool = False,
-        into_series_store: Optional[str] = None,
-        leave_case_out: bool = False,
-        new_case_threshold: Literal["max", "min", "most_similar"] = "min",
-        num_cases_to_generate: Optional[int] = None,
-        ordered_by_specified_features: bool = False,
-        post_process_features: Optional[Iterable[str]] = None,
-        post_process_values: Optional[List[List[object]]] = None,
-        preserve_feature_values: Optional[Iterable[str]] = None,
-        substitute_output: bool = True,
-        use_case_weights: bool = False,
-        use_regional_model_residuals: bool = True,
-        weight_feature: Optional[str] = None
-    ) -> Tuple[Dict, int, int]:
-        """
-        Multiple case react.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        context_values : list of list of object, optional
-            A 2d list of context values to react to.
-            If None for discriminative react, it is assumed that `session`
-            and `session_id` keys are set in the `details`.
-        action_features : iterable of str, optional
-            An iterable of feature names to treat as action features during
-            react.
-        action_values : list of list of object, optional
-            One or more action values to use for action features.
-            If specified, will only return the specified explanation
-            details for the given actions. (Discriminative reacts only)
-        allow_nulls : bool, default False
-            When true will allow return of null values if there
-            are nulls in the local model for the action features, applicable
-            only to discriminative reacts.
-        context_features : iterable of str, optional
-            An iterable of feature names to treat as context features during
-            react.
-        derived_context_features : iterable of str, optional
-            An iterable of feature names whose values should be computed
-            from the provided context in the specified order. Must be different
-            than context_features.
-        derived_action_features : iterable of str, optional
-            An iterable of feature names whose values should be computed
-            after generation from the generated case prior to output, in the
-            specified order. Must be a subset of action_features.
-        exclude_novel_nominals_from_uniqueness_check : bool, default False
-            If True, will exclude features which have a subtype defined in their feature
-            attributes from the uniqueness check that happens when ``generate_new_cases``
-            is True. Only applies to generative reacts.
-        input_is_substituted : bool, default False
-            if True assumes provided categorical (nominal or
-            ordinal) feature values have already been substituted.
-        substitute_output : bool, default True
-            If False, will not substitute categorical feature
-            values. Only applicable if a substitution value map has been set.
-        details : dict, optional
-            If details are specified, the response will
-            contain the requested explanation data along with the reaction.
-        desired_conviction : float
-            If specified will execute a generative react. If not
-            specified will executed a discriminative react. Conviction is the
-            ratio of expected surprisal to generated surprisal for each
-            feature generated, valid values are in the range of zero to
-            infinity.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each
-            case's weight_feature weight.
-        case_indices : Iterable of Sequence[Union[str, int]], defaults to None
-            An Iterable of Sequences, of session id and index, where
-            index is the original 0-based index of the case as it was trained
-            into the session. If this case does not exist, discriminative react
-            outputs null, generative react ignores it.
-        post_process_features : iterable of str, optional
-            List of feature names that will be made available during the
-            execution of post_process feature attributes.
-        post_process_values : list of list of object, optional
-            A 2d list of values corresponding to post_process_features that
-            will be made available during the execution of post_process feature
-            attributes.
-        preserve_feature_values : iterable of str
-            List of features that will preserve their values from
-            the case specified by case_indices, appending and overwriting the
-            specified contexts as necessary.  For generative reacts, if
-            case_indices isn't specified will preserve feature values of a
-            random case.
-        leave_case_out : bool, default False
-            If set to True and specified along with case_indices,
-            each individual react will respectively ignore the corresponding
-            case specified by case_indices by leaving it out.
-        into_series_store : str, optional
-            The name of a series store. If specified, will store an internal
-            record of all react contexts for this session and series to be used
-            later with train series.
-        use_regional_model_residuals : bool
-            If false uses model feature residuals, if True
-            recalculates regional model residuals.
-        feature_bounds_map : dict of dict
-            A mapping of feature names to the bounds for the
-            feature values to be generated in.
-        generate_new_cases : {"always", "attempt", "no"}, default "no"
-            How to generate new cases.
-        ordered_by_specified_features : bool, default False
-            If True order of generated feature values will match
-            the order of specified features.
-        num_cases_to_generate : int, default 1
-            The number of cases to generate.
-        new_case_threshold : {"min", "max", "most_similar"}, optional
-            Distance to determine the privacy cutoff. If None,
-            will default to "min".
-
-        Returns
-        -------
-        dict
-            The react result including audit details.
-        int
-            The request payload size.
-        int
-            The result payload size.
-        """
-        return self._execute_sized(trainee_id, "batch_react", {
-            "context_features": context_features,
-            "context_values": context_values,
-            "action_features": action_features,
-            "action_values": action_values,
-            "derived_context_features": derived_context_features,
-            "derived_action_features": derived_action_features,
-            "details": details,
-            "exclude_novel_nominals_from_uniqueness_check": exclude_novel_nominals_from_uniqueness_check,
-            "extra_features": extra_features,
-            "case_indices": case_indices,
-            "allow_nulls": allow_nulls,
-            "input_is_substituted": input_is_substituted,
-            "substitute_output": substitute_output,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-            "leave_case_out": leave_case_out,
-            "num_cases_to_generate": num_cases_to_generate,
-            "use_regional_model_residuals": use_regional_model_residuals,
-            "desired_conviction": desired_conviction,
-            "feature_bounds_map": feature_bounds_map,
-            "generate_new_cases": generate_new_cases,
-            "ordered_by_specified_features": ordered_by_specified_features,
-            "post_process_features": post_process_features,
-            "post_process_values": post_process_values,
-            "preserve_feature_values": preserve_feature_values,
-            "new_case_threshold": new_case_threshold,
-            "into_series_store": into_series_store,
-        })
-
-    def batch_react_series(
-        self,
-        trainee_id: str,
-        *,
-        action_features: Optional[Iterable[str]] = None,
-        action_values: Optional[List[List[object]]] = None,
-        case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        context_values: Optional[List[List[object]]] = None,
-        context_features: Optional[Iterable[str]] = None,
-        continue_series: Optional[bool] = False,
-        continue_series_features: Optional[Iterable[str]] = None,
-        continue_series_values: Optional[Union[List[object], List[List[object]]]] = None,
-        derived_action_features: Optional[Iterable[str]] = None,
-        derived_context_features: Optional[Iterable[str]] = None,
-        desired_conviction: Optional[float] = None,
-        details: Optional[Dict] = None,
-        exclude_novel_nominals_from_uniqueness_check: bool = False,
-        extra_features: Optional[Iterable[str]] = None,
-        feature_bounds_map: Optional[Dict] = None,
-        final_time_steps: Optional[Union[List[object], List[List[object]]]] = None,
-        generate_new_cases: Literal["always", "attempt", "no"] = "no",
-        init_time_steps: Optional[Union[List[object], List[List[object]]]] = None,
-        initial_features: Optional[Iterable[str]] = None,
-        initial_values: Optional[Union[List[object], List[List[object]]]] = None,
-        input_is_substituted: bool = False,
-        leave_case_out: bool = False,
-        max_series_lengths: Optional[List[int]] = None,
-        new_case_threshold: Literal["max", "min", "most_similar"] = "min",
-        num_series_to_generate: int = 1,
-        ordered_by_specified_features: bool = False,
-        output_new_series_ids: bool = True,
-        preserve_feature_values: Optional[Iterable[str]] = None,
-        series_context_features: Optional[Iterable[str]] = None,
-        series_context_values: Optional[Union[List[object], List[List[object]]]] = None,
-        series_id_tracking: Literal["dynamic", "fixed", "no"] = "fixed",
-        series_stop_maps: Optional[List[Dict[str, Dict]]] = None,
-        substitute_output: bool = True,
-        use_case_weights: bool = False,
-        use_regional_model_residuals: bool = True,
-        weight_feature: Optional[str] = None
-    ) -> Tuple[Dict, int, int]:
-        """
-        React in a series until a series_stop_map condition is met.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        num_series_to_generate : int, optional
-            The number of series to generate.
-        final_time_steps : list of object, optional
-            The time steps at which to end synthesis. Time-series
-            only. Must provide either one for all series, or exactly one per
-            series.
-        init_time_steps : list of object, optional
-            The time steps at which to begin synthesis. Time-series
-            only. Must provide either one for all series, or exactly one per
-            series.
-        initial_features : iterable of str, optional
-            List of features to condition just the first case in a
-            series, overwrites context_features and derived_context_features
-            for that first case. All specified initial features must be in one
-            of: context_features, action_features, derived_context_features or
-            derived_action_features. If provided a value that isn't in one of
-            those lists, it will be ignored.
-        initial_values : list of list of object, optional
-            2d list of values corresponding to the initial_features,
-            used to condition just the first case in each series. Must provide
-            either one for all series, or exactly one per series.
-        series_stop_maps : list of dict of dict, optional
-            A dictionary of feature name to stop conditions. Must provide either
-            one for all series, or exactly one per series.
-        max_series_lengths : list of int, optional
-            maximum size a series is allowed to be.  Default is
-            3 * model_size, a 0 or less is no limit. If forecasting
-            with ``continue_series``, this defines the maximum length of the
-            forecast. Must provide either one for all series, or exactly
-            one per series.
-        continue_series : bool, default False
-            When True will attempt to continue existing series instead of
-            starting new series. If `initial_values` provide series IDs, it
-            will continue those explicitly specified IDs, otherwise it will
-            randomly select series to continue.
-            .. note::
-
-                Terminated series with terminators cannot be continued and
-                will result in null output.
-        continue_series_features : list of str, optional
-            The list of feature names corresponding to the values in each row of
-            `continue_series_values`. This value is ignored if
-            `continue_series_values` is None.
-        continue_series_values : list of list of list of object or list of pandas.DataFrame, default None
-            The set of series data to be forecasted with feature values in the
-            same order defined by `continue_series_values`. The value of
-            `continue_series` will be ignored and treated as true if this value
-            is specified.
-        derived_context_features : iterable of str, optional
-            List of context features whose values should be computed
-            from the entire series in the specified order. Must be
-            different than context_features.
-        derived_action_features : iterable of str, optional
-            List of action features whose values should be computed
-            from the resulting last row in series, in the specified order.
-            Must be a subset of action_features.
-        series_context_features : iterable of str, optional
-            List of context features corresponding to
-            series_context_values, if specified must not overlap with any
-            initial_features or context_features.
-        series_context_values : list of list of list of object or list, optional
-            3d-list of context values, one for each feature for each
-            row for each series. If specified, max_series_lengths are ignored.
-        output_new_series_ids : bool, default True
-            If True, series ids are replaced with unique values on output.
-            If False, will maintain or replace ids with existing trained values,
-            but also allows output of series with duplicate existing ids.
-        series_id_tracking : {"dynamic", "fixed", "no"}, default "fixed"
-            Controls how closely generated series should follow existing series.
-        context_values: list of list of object
-            See parameter ``contexts`` in :meth:`react`.
-        action_features: iterable of str
-            See parameter ``action_features`` in :meth:`react`.
-        action_values: list of list of object
-            See parameter ``actions`` in :meth:`react`.
-        context_features: iterable of str
-            See parameter ``context_features`` in :meth:`react`.
-        input_is_substituted : bool, default False
-            See parameter ``input_is_substituted`` in :meth:`react`.
-        substitute_output : bool
-            See parameter ``substitute_output`` in :meth:`react`.
-        details: dict, optional
-            See parameter ``details`` in :meth:`react`.
-        desired_conviction: float
-            See parameter ``desired_conviction`` in :meth:`react`.
-        exclude_novel_nominals_from_uniqueness_check : bool, default False
-            See parameter ``exclude_novel_nominals_from_uniqueness_check`` in :meth:`react`.
-        weight_feature : str
-            See parameter ``weight_feature`` in :meth:`react`.
-        use_case_weights : bool
-            See parameter ``use_case_weights`` in :meth:`react`.
-        case_indices: iterable of sequence of str, int
-            See parameter ``case_indices`` in :meth:`react`.
-        preserve_feature_values : iterable of str
-            See parameter ``preserve_feature_values`` in :meth:`react`.
-        new_case_threshold : str
-            See parameter ``new_case_threshold`` in :meth:`react`.
-        leave_case_out : bool
-            See parameter ``leave_case_out`` in :meth:`react`.
-        use_regional_model_residuals : bool
-            See parameter ``use_regional_model_residuals`` in :meth:`react`.
-        feature_bounds_map: dict of dict
-            See parameter ``feature_bounds_map`` in :meth:`react`.
-        generate_new_cases : {"always", "attempt", "no"}
-            See parameter ``generate_new_cases`` in :meth:`react`.
-        ordered_by_specified_features : bool
-            See parameter ``ordered_by_specified_features`` in :meth:`react`.
-
-        Returns
-        -------
-        dict
-            A dictionary with keys `action_features` and `series`. Where
-            `series` is a 2d list of values (rows of data per series), and
-            `action_features` is the list of all action features
-            (specified and derived).
-        int
-            The request payload size.
-        int
-            The result payload size.
-        """
-        return self._execute_sized(trainee_id, "batch_react_series", {
-            "context_features": context_features,
-            "context_values": context_values,
-            "action_features": action_features,
-            "action_values": action_values,
-            "final_time_steps": final_time_steps,
-            "init_time_steps": init_time_steps,
-            "initial_features": initial_features,
-            "initial_values": initial_values,
-            "series_stop_maps": series_stop_maps,
-            "max_series_lengths": max_series_lengths,
-            "continue_series": continue_series,
-            "continue_series_features": continue_series_features,
-            "continue_series_values": continue_series_values,
-            "derived_context_features": derived_context_features,
-            "derived_action_features": derived_action_features,
-            "series_context_features": series_context_features,
-            "series_context_values": series_context_values,
-            "series_id_tracking": series_id_tracking,
-            "output_new_series_ids": output_new_series_ids,
-            "details": details,
-            "exclude_novel_nominals_from_uniqueness_check": exclude_novel_nominals_from_uniqueness_check,
-            "extra_features": extra_features,
-            "case_indices": case_indices,
-            "input_is_substituted": input_is_substituted,
-            "substitute_output": substitute_output,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-            "leave_case_out": leave_case_out,
-            "num_series_to_generate": num_series_to_generate,
-            "preserve_feature_values": preserve_feature_values,
-            "new_case_threshold": new_case_threshold,
-            "use_regional_model_residuals": use_regional_model_residuals,
-            "desired_conviction": desired_conviction,
-            "feature_bounds_map": feature_bounds_map,
-            "generate_new_cases": generate_new_cases,
-            "ordered_by_specified_features": ordered_by_specified_features,
-        })
-
-    def react_into_features(
-        self,
-        trainee_id: str,
-        *,
-        distance_contribution: bool = False,
-        familiarity_conviction_addition: bool = False,
-        familiarity_conviction_removal: bool = False,
-        features: Optional[Iterable[str]] = None,
-        influence_weight_entropy: Union[bool, str] = False,
-        p_value_of_addition: bool = False,
-        p_value_of_removal: bool = False,
-        similarity_conviction: bool = False,
-        use_case_weights: bool = False,
-        weight_feature: Optional[str] = None,
-    ) -> None:
-        """
-        Calculate and cache conviction and other statistics.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        features : iterable of str, optional
-            An iterable of features to calculate convictions.
-        familiarity_conviction_addition : bool or str, default False
-            The name of the feature to store conviction of addition
-            values. If set to True the values will be stored to the feature
-            'familiarity_conviction_addition'.
-        familiarity_conviction_removal : bool or str, default False
-            The name of the feature to store conviction of removal
-            values. If set to True the values will be stored to the feature
-            'familiarity_conviction_removal'.
-        influence_weight_entropy : bool or str, default False
-            The name of the feature to store influence weight entropy values in.
-            If set to True, the values will be stored in the feature
-            'influence_weight_entropy'.
-        p_value_of_addition : bool or str, default False
-            The name of the feature to store p value of addition
-            values. If set to True the values will be stored to the feature
-            'p_value_of_addition'.
-        p_value_of_removal : bool or str, default False
-            The name of the feature to store p value of removal
-            values. If set to True the values will be stored to the feature
-            'p_value_of_removal'.
-        similarity_conviction : bool or str, default False
-            The name of the feature to store similarity conviction
-            values. If set to True the values will be stored to the feature
-            'similarity_conviction'.
-        distance_contribution : bool or str, default False
-            The name of the feature to store distance contribution.
-            If set to True the values will be stored to the
-            feature 'distance_contribution'.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each
-            case's weight_feature weight.
-        """
-        return self._execute(trainee_id, "react_into_features", {
-            "features": features,
-            "familiarity_conviction_addition": familiarity_conviction_addition,
-            "familiarity_conviction_removal": familiarity_conviction_removal,
-            "influence_weight_entropy": influence_weight_entropy,
-            "p_value_of_addition": p_value_of_addition,
-            "p_value_of_removal": p_value_of_removal,
-            "similarity_conviction": similarity_conviction,
-            "distance_contribution": distance_contribution,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-        })
-
-    def batch_react_group(
-        self,
-        trainee_id: str,
-        new_cases: List[List[List[object]]],
-        *,
-        features: Optional[Iterable[str]] = None,
-        distance_contributions: bool = False,
-        familiarity_conviction_addition: bool = True,
-        familiarity_conviction_removal: bool = False,
-        kl_divergence_addition: bool = False,
-        kl_divergence_removal: bool = False,
-        p_value_of_addition: bool = False,
-        p_value_of_removal: bool = False,
-        weight_feature: Optional[str] = None,
-        use_case_weights: bool = False
-    ) -> Dict:
-        """
-        Computes specified data for a **set** of cases.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        new_cases : list of list of list of object or list
-            Specify a **set** using a list of cases to compute the conviction of
-            groups of cases as shown in the following example.
-        features : iterable of str, optional
-            An iterable of feature names to consider while calculating
-            convictions.
-        distance_contributions : bool, default False
-            Calculate and output distance contribution ratios in
-            the output dict for each case.
-        familiarity_conviction_addition : bool, default True
-            Calculate and output familiarity conviction of adding the
-            specified cases.
-        familiarity_conviction_removal : bool, default False
-            Calculate and output familiarity conviction of removing
-            the specified cases.s
-        kl_divergence_addition : bool, default False
-            Calculate and output KL divergence of adding the
-            specified cases.
-        kl_divergence_removal : bool, default False
-            Calculate and output KL divergence of removing the
-            specified cases.
-        p_value_of_addition : bool, default False
-            If true will output p value of addition.
-        p_value_of_removal : bool, default False
-            If true will output p value of removal.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each
-            case's weight_feature weight.
-
-        Returns
-        -------
-        dict
-            The react response.
-        """
-        return self._execute(trainee_id, "batch_react_group", {
-            "features": features,
-            "new_cases": new_cases,
-            "distance_contributions": distance_contributions,
-            "familiarity_conviction_addition": familiarity_conviction_addition,
-            "familiarity_conviction_removal": familiarity_conviction_removal,
-            "kl_divergence_addition": kl_divergence_addition,
-            "kl_divergence_removal": kl_divergence_removal,
-            "p_value_of_addition": p_value_of_addition,
-            "p_value_of_removal": p_value_of_removal,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-        })
-
-    def react_aggregate(
-        self,
-        trainee_id: str,
-        *,
-        action_feature: Optional[str] = None,
-        confusion_matrix_min_count: Optional[int] = None,
-        context_features: Optional[Iterable[str]] = None,
-        details: Optional[dict] = None,
-        feature_influences_action_feature: Optional[str] = None,
-        hyperparameter_param_path: Optional[Iterable[str]] = None,
-        num_robust_influence_samples: Optional[int] = None,
-        num_robust_residual_samples: Optional[int] = None,
-        num_robust_influence_samples_per_case: Optional[int] = None,
-        num_samples: Optional[int] = None,
-        prediction_stats_action_feature: Optional[str] = None,
-        residuals_hyperparameter_feature: Optional[str] = None,
-        robust_hyperparameters: Optional[bool] = None,
-        sample_model_fraction: Optional[float] = None,
-        sub_model_size: Optional[int] = None,
-        use_case_weights: bool = False,
-        weight_feature: Optional[str] = None,
-    ) -> dict:
-        """
-        Reacts into the aggregate trained cases in the Trainee.
-
-        Calculates, caches, and/or returns the requested influences and prediction stats.
-
-        Parameters
-        ----------
-        action_feature : str, optional
-            Name of target feature for which to do computations. If ``prediction_stats_action_feature``
-            and ``feature_influences_action_feature`` are not provided, they will default to this value.
-            If ``feature_influences_action_feature`` is not provided and feature influences ``details`` are
-            selected, this feature must be provided.
-        confusion_matrix_min_count : int, optional
-            The number of predictions a class should have (value of a cell in the
-            matrix) for it to remain in the confusion matrix. If the count is
-            less than this value, it will be accumulated into a single value of
-            all insignificant predictions for the class and removed from the
-            confusion matrix. Defaults to 10, applicable only to confusion
-            matrices when computing residuals.
-        context_features : iterable of str, optional
-            List of features names to use as contexts for
-            computations. Default is all trained non-unique features if
-            unspecified.
-        details : map of str -> object, optional
-            If details are specified, the response will contain the requested
-            explanation data.. Below are the valid keys and data types for the
-            different audit details. Omitted keys, values set to None, or False
-            values for Booleans will not be included in the data returned.
-
-            - prediction_stats : bool, optional
-                If True outputs full feature prediction stats for all (context and action) features.
-                The prediction stats returned are set by the "selected_prediction_stats" parameter
-                in the `details` parameter. Uses full calculations, which uses leave-one-out for
-                features for computations. False removes cached values.
-            - feature_residuals_full : bool, optional
-                For each context_feature, use the full set of all other context_features to predict
-                the feature. False removes cached values. When ``prediction_stats``
-                in the ``details`` parameter is true, the Trainee will also calculate and cache the
-                full feature residuals.
-            - feature_residuals_robust : bool, optional
-                For each context_feature, use the robust (power set/permutations) set of all other
-                context_features to predict the feature. False removes cached values.
-            - feature_contributions_full : bool, optional
-                For each context_feature, use the full set of all other
-                context_features to compute the mean absolute delta between
-                prediction of action feature with and without the context features
-                in the model. False removes cached values.
-            - feature_contributions_robust : bool, optional
-                For each context_feature, use the robust (power set/permutation)
-                set of all other context_features to compute the mean absolute
-                delta between prediction of the action feature with and without the
-                context features in the model. False removes cached values.
-            - feature_mda_full : bool, optional
-                When True will compute Mean Decrease in Accuracy (MDA)
-                for each context feature at predicting the action feature. Drop
-                each feature and use the full set of remaining context features
-                for each prediction. False removes cached values.
-            - feature_mda_robust : bool, optional
-                Compute Mean Decrease in Accuracy MDA by dropping each feature and using the
-                robust (power set/permutations) set of remaining context features
-                for each prediction. False removes cached values.
-            - feature_feature_mda_permutation_full : bool, optional
-                Compute MDA by scrambling each feature and using the
-                full set of remaining context features for each prediction.
-                False removes cached values.
-            - feature_feature_mda_permutation_robust : bool, optional
-                Compute MDA by scrambling each feature and using the
-                robust (power set/permutations) set of remaining context features
-                for each prediction. False removes cached values.
-            - action_condition : map of str -> any, optional
-                A condition map to select the action set, which is the dataset for which
-                the prediction stats are for. If both ``action_condition`` and ``context_condition``
-                are provided, then all of the action cases selected by the ``action_condition``
-                will be excluded from the context set, which is the set being queried to make to
-                make predictions on the action set, effectively holding them out.
-                If only ``action_condition`` is specified, then only the single predicted case
-                will be left out.
-
-                .. NOTE::
-                    The dictionary keys are the feature name and values are one of:
-
-                        - None
-                        - A value, must match exactly.
-                        - An array of two numeric values, specifying an inclusive
-                        range. Only applicable to continuous and numeric ordinal
-                        features.
-                        - An array of string values, must match any of these values
-                        exactly. Only applicable to nominal and string ordinal
-                        features.
-            - action_num_cases : int, optional
-                The maximum amount of cases to use to calculate prediction stats.
-                If not specified, the limit will be k cases if precision is
-                "similar", or 1000 cases if precision is "exact". Works with or
-                without ``action_condition``.
-                -If ``action_condition`` is set:
-                    If None, will be set to k if precision is "similar" or no limit if precision is "exact".
-                - If ``action_condition`` is not set:
-                    If None, will be set to the Howso default limit of 2000.
-            - action_condition_precision : {"exact", "similar"}, optional
-                The precision to use when selecting cases with the ``action_condition``.
-                If not specified "exact" will be used. Only used if ``action_condition``
-                is not None.
-            - context_condition : map of str -> any, optional
-                A condition map to select the context set, which is the set being queried to make
-                to make predictions on the action set. If both ``action_condition`` and ``context_condition``
-                are provided,  then all of the cases from the action set, which is the dataset for which the
-                prediction stats are for, will be excluded from the context set, effectively holding them out.
-                If only ``action_condition`` is specified,  then only the single predicted case will be left out.
-
-                .. NOTE::
-                    The dictionary keys are the feature name and values are one of:
-
-                        - None
-                        - A value, must match exactly.
-                        - An array of two numeric values, specifying an inclusive
-                        range. Only applicable to continuous and numeric ordinal
-                        features.
-                        - An array of string values, must match any of these values
-                        exactly. Only applicable to nominal and string ordinal
-                        features.
-            - context_precision_num_cases : int, optional
-                Limit on the number of context cases when ``context_condition_precision`` is set to "similar".
-                If None, will be set to k.
-            - context_condition_precision : {"exact", "similar"}, optional
-                The precision to use when selecting cases with the ``context_condition``.
-                If not specified "exact" will be used. Only used if ``context_condition``
-                is not None.
-            - prediction_stats_features : list, optional
-                List of features to use when calculating conditional prediction stats. Should contain all
-                action and context features desired. If ``action_feature`` is also provided, that feature will
-                automatically be appended to this list if it is not already in the list.
-                    stats : list of str, optional
-            - missing_value_accuracy_full : bool, optional
-                The number of cases with missing values predicted to have missing values divided by the number
-                of cases with missing values, applies to all features that contain missing values. Uses full
-                calculations.
-            - missing_value_accuracy_robust : bool, optional
-                The number of cases with missing values predicted to have missing values divided by the number
-                of cases with missing values, applies to all features that contain missing values. Uses robust
-                calculations.
-            - selected_prediction_stats : list, optional
-                List of stats to output. When unspecified, returns all except the confusion matrix. Allowed values:
-
-                - all : Returns all the the available prediction stats, including the confusion matrix.
-                - accuracy : The number of correct predictions divided by the
-                  total number of predictions.
-                - confusion_matrix : A sparse map of actual feature value to a map of
-                  predicted feature value to counts.
-                - mae : Mean absolute error. For continuous features, this is
-                  calculated as the mean of absolute values of the difference
-                  between the actual and predicted values. For nominal features,
-                  this is 1 - the average categorical action probability of each case's
-                  correct classes. Categorical action probabilities are the probabilities
-                  for each class for the action feature.
-                - mda : Mean decrease in accuracy when each feature is dropped
-                  from the model, applies to all features.
-                - feature_mda_permutation_full : Mean decrease in accuracy that used
-                  scrambling of feature values instead of dropping each
-                  feature, applies to all features.
-                - precision : Precision (positive predictive) value for nominal
-                  features only.
-                - r2 : The r-squared coefficient of determination, for
-                  continuous features only.
-                - recall : Recall (sensitivity) value for nominal features only.
-                - rmse : Root mean squared error, for continuous features only.
-                - spearman_coeff : Spearman's rank correlation coefficient,
-                  for continuous features only.
-                - mcc : Matthews correlation coefficient, for nominal features only.
-        feature_influences_action_feature : str, optional
-            When feature influences such as contributions and mda, use this feature as
-            the action feature.  If not provided, will default to the ``action_feature`` if provided.
-            If ``action_feature`` is not provided and feature influences ``details`` are
-            selected, this feature must be provided.
-        hyperparameter_param_path : iterable of str, optional.
-            Full path for hyperparameters to use for computation. If specified
-            for any residual computations, takes precedence over action_feature
-            parameter.  Can be set to a 'paramPath' value from the results of
-            'get_params()' for a specific set of hyperparameters.
-        num_robust_influence_samples : int, optional
-            Total sample size of model to use (using sampling with replacement)
-            for robust contribution computation. Defaults to 300.
-        num_robust_residual_samples : int, optional
-            Total sample size of model to use (using sampling with replacement)
-            for robust mda and residual computation.
-            Defaults to 1000 * (1 + log(number of features)).  Note: robust mda
-            will be updated to use num_robust_influence_samples in a future release.
-        num_robust_influence_samples_per_case : int, optional
-            Specifies the number of robust samples to use for each case for
-            robust contribution computations.
-            Defaults to 300 + 2 * (number of features).
-        num_samples : int, optional
-            Total sample size of model to use (using sampling with replacement)
-            for all non-robust computation. Defaults to 1000.
-            If specified overrides sample_model_fraction.```
-        residuals_hyperparameter_feature : str, optional
-            When calculating residuals and prediction stats, uses this target
-            features's hyperparameters. The trainee must have been analyzed with
-            this feature as the action feature first. If not provided, by default
-            residuals and prediction stats uses ".targetless" hyperparameters.
-        robust_hyperparameters : bool, optional
-            When specified, will attempt to return residuals that were
-            computed using hyperparameters with the specified robust or
-            non-robust type.
-        prediction_stats_action_feature : str, optional
-            When calculating residuals and prediction stats, uses this target features's
-            hyperparameters. The trainee must have been analyzed with this feature as the
-            action feature first. If both ``prediction_stats_action_feature`` and
-            ``action_feature`` are not provided, by default residuals and prediction
-            stats uses ".targetless" hyperparameters. If "action_feature" is provided,
-            and this value is not provided, will default to ``action_feature``.
-        sample_model_fraction : float, optional
-            A value between 0.0 - 1.0, percent of model to use in sampling
-            (using sampling without replacement). Applicable only to non-robust
-            computation. Ignored if num_samples is specified.
-            Higher values provide better accuracy at the cost of compute time.
-        sub_model_size : int, optional
-            Subset of model to use for calculations. Applicable only
-            to models > 1000 cases.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each case's
-            weight_feature weight.
-        weight_feature : str, optional
-            The name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-
-        Returns
-        -------
-        dict of str to dict of str to float
-            If specified, a map of feature to map of stat type to stat values is returned.
-        """
-        return self._execute(trainee_id, "react_aggregate", {
-            "action_feature": action_feature,
-            "residuals_hyperparameter_feature": residuals_hyperparameter_feature,
-            "context_features": context_features,
-            "confusion_matrix_min_count": confusion_matrix_min_count,
-            "details": details,
-            "feature_influences_action_feature": feature_influences_action_feature,
-            "hyperparameter_param_path": hyperparameter_param_path,
-            "num_robust_influence_samples": num_robust_influence_samples,
-            "num_robust_residual_samples": num_robust_residual_samples,
-            "num_robust_influence_samples_per_case": num_robust_influence_samples_per_case,
-            "num_samples": num_samples,
-            "prediction_stats_action_feature": prediction_stats_action_feature,
-            "robust_hyperparameters": robust_hyperparameters,
-            "sample_model_fraction": sample_model_fraction,
-            "sub_model_size": sub_model_size,
-            "use_case_weights": use_case_weights,
-            "weight_feature": weight_feature,
         })
 
     def get_session_indices(self, trainee_id: str, session: str) -> List[int]:
